adding html example about create an s100 api
