+++ s100py/docs/html/s100py.html
@@ -18,7 +18,7 @@
     
     <link rel="index" title="Index" href="genindex.html" />
     <link rel="search" title="Search" href="search.html" />
-    <link rel="next" title="Extending the API" href="extending_the_api.html" /> 
+    <link rel="next" title="s100py.s1xx Module" href="s1xx.html" /> 
   </head><body>
     <div class="related" role="navigation" aria-label="related navigation">
       <h3>Navigation</h3>
@@ -30,7 +30,7 @@
           <a href="py-modindex.html" title="Python Module Index"
              >modules</a> |</li>
         <li class="right" >
-          <a href="extending_the_api.html" title="Extending the API"
+          <a href="s1xx.html" title="s100py.s1xx Module"
              accesskey="N">next</a> |</li>
         <li class="nav-item nav-item-0"><a href="#">Pydro  documentation</a> &#187;</li> 
       </ul>
@@ -134,7 +134,6 @@ so we should have initialized first.</p>
 </div>
 <div class="toctree-wrapper compound">
 <ul>
-<li class="toctree-l1"><a class="reference internal" href="extending_the_api.html">Extending the API</a></li>
 <li class="toctree-l1"><a class="reference internal" href="s1xx.html">s100py.s1xx Module</a><ul>
 <li class="toctree-l2"><a class="reference internal" href="s1xx.html#classes">Classes</a><ul>
 <li class="toctree-l3"><a class="reference internal" href="api/s100py.s1xx.S1XXFile.html">S1XXFile</a></li>
@@ -152,6 +151,7 @@ so we should have initialized first.</p>
 <li class="toctree-l3"><a class="reference internal" href="api/s100py.s100.COMMON_POINT_RULE.html">COMMON_POINT_RULE</a></li>
 <li class="toctree-l3"><a class="reference internal" href="api/s100py.s100.DATA_CODING_FORMAT.html">DATA_CODING_FORMAT</a></li>
 <li class="toctree-l3"><a class="reference internal" href="api/s100py.s100.DirectPosition.html">DirectPosition</a></li>
+<li class="toctree-l3"><a class="reference internal" href="api/s100py.s100.FeatureInformation.html">FeatureInformation</a></li>
 <li class="toctree-l3"><a class="reference internal" href="api/s100py.s100.FeatureInstance.html">FeatureInstance</a></li>
 <li class="toctree-l3"><a class="reference internal" href="api/s100py.s100.FeatureInstanceTimeSeries.html">FeatureInstanceTimeSeries</a></li>
 <li class="toctree-l3"><a class="reference internal" href="api/s100py.s100.FeatureInstance_Format_2.html">FeatureInstance_Format_2</a></li>
@@ -181,12 +181,9 @@ so we should have initialized first.</p>
 <li class="toctree-l3"><a class="reference internal" href="api/s100py.s102.BathymetryCoverages_List.html">BathymetryCoverages_List</a></li>
 <li class="toctree-l3"><a class="reference internal" href="api/s100py.s102.BathymetryFeatureInstance.html">BathymetryFeatureInstance</a></li>
 <li class="toctree-l3"><a class="reference internal" href="api/s100py.s102.BathymetryGroup_List.html">BathymetryGroup_List</a></li>
-<li class="toctree-l3"><a class="reference internal" href="api/s100py.s102.BathymetryValueRecord.html">BathymetryValueRecord</a></li>
 <li class="toctree-l3"><a class="reference internal" href="api/s100py.s102.BathymetryValues.html">BathymetryValues</a></li>
-<li class="toctree-l3"><a class="reference internal" href="api/s100py.s102.BathymetryValuesList.html">BathymetryValuesList</a></li>
 <li class="toctree-l3"><a class="reference internal" href="api/s100py.s102.DiscoveryMetadata.html">DiscoveryMetadata</a></li>
 <li class="toctree-l3"><a class="reference internal" href="api/s100py.s102.FeatureCodes.html">FeatureCodes</a></li>
-<li class="toctree-l3"><a class="reference internal" href="api/s100py.s102.FeatureInformation.html">FeatureInformation</a></li>
 <li class="toctree-l3"><a class="reference internal" href="api/s100py.s102.FeatureInformation_dataset.html">FeatureInformation_dataset</a></li>
 <li class="toctree-l3"><a class="reference internal" href="api/s100py.s102.S102File.html">S102File</a></li>
 <li class="toctree-l3"><a class="reference internal" href="api/s100py.s102.S102Root.html">S102Root</a></li>
@@ -206,6 +203,8 @@ so we should have initialized first.</p>
 <li class="toctree-l2"><a class="reference internal" href="s102.html#class-inheritance-diagram">Class Inheritance Diagram</a></li>
 </ul>
 </li>
+<li class="toctree-l1"><a class="reference internal" href="extending_the_api.html">Extending the API</a></li>
+<li class="toctree-l1"><a class="reference internal" href="sample_api.html">An Example API</a></li>
 </ul>
 </div>
 </div>
@@ -218,15 +217,16 @@ so we should have initialized first.</p>
         <div class="sphinxsidebarwrapper">
   <h3><a href="#">Table of Contents</a></h3>
   <ul>
-<li class="toctree-l1"><a class="reference internal" href="extending_the_api.html">Extending the API</a></li>
 <li class="toctree-l1"><a class="reference internal" href="s1xx.html">s100py.s1xx Module</a></li>
 <li class="toctree-l1"><a class="reference internal" href="s100.html">s100py.s100 Module</a></li>
 <li class="toctree-l1"><a class="reference internal" href="s102.html">s100py.s102 Module</a></li>
+<li class="toctree-l1"><a class="reference internal" href="extending_the_api.html">Extending the API</a></li>
+<li class="toctree-l1"><a class="reference internal" href="sample_api.html">An Example API</a></li>
 </ul>
 
   <h4>Next topic</h4>
-  <p class="topless"><a href="extending_the_api.html"
-                        title="next chapter">Extending the API</a></p>
+  <p class="topless"><a href="s1xx.html"
+                        title="next chapter">s100py.s1xx Module</a></p>
 <div id="searchbox" style="display: none" role="search">
   <h3 id="searchlabel">Quick search</h3>
     <div class="searchformwrapper">
@@ -251,7 +251,7 @@ so we should have initialized first.</p>
           <a href="py-modindex.html" title="Python Module Index"
              >modules</a> |</li>
         <li class="right" >
-          <a href="extending_the_api.html" title="Extending the API"
+          <a href="s1xx.html" title="s100py.s1xx Module"
              >next</a> |</li>
         <li class="nav-item nav-item-0"><a href="#">Pydro  documentation</a> &#187;</li> 
       </ul>
