+++ s100py/docs/html/s1xx.html
@@ -19,7 +19,7 @@
     <link rel="index" title="Index" href="genindex.html" />
     <link rel="search" title="Search" href="search.html" />
     <link rel="next" title="S1XXFile" href="api/s100py.s1xx.S1XXFile.html" />
-    <link rel="prev" title="Extending the API" href="extending_the_api.html" /> 
+    <link rel="prev" title="s100py" href="s100py.html" /> 
   </head><body>
     <div class="related" role="navigation" aria-label="related navigation">
       <h3>Navigation</h3>
@@ -34,7 +34,7 @@
           <a href="api/s100py.s1xx.S1XXFile.html" title="S1XXFile"
              accesskey="N">next</a> |</li>
         <li class="right" >
-          <a href="extending_the_api.html" title="Extending the API"
+          <a href="s100py.html" title="s100py"
              accesskey="P">previous</a> |</li>
         <li class="nav-item nav-item-0"><a href="s100py.html">Pydro  documentation</a> &#187;</li> 
       </ul>
@@ -114,7 +114,6 @@ implementing one of the S100 family of specifications.</p>
         <div class="sphinxsidebarwrapper">
   <h3><a href="s100py.html">Table of Contents</a></h3>
   <ul class="current">
-<li class="toctree-l1"><a class="reference internal" href="extending_the_api.html">Extending the API</a></li>
 <li class="toctree-l1 current"><a class="current reference internal" href="#">s100py.s1xx Module</a><ul>
 <li class="toctree-l2"><a class="reference internal" href="#classes">Classes</a><ul>
 <li class="toctree-l3"><a class="reference internal" href="api/s100py.s1xx.S1XXFile.html">S1XXFile</a></li>
@@ -129,11 +128,13 @@ implementing one of the S100 family of specifications.</p>
 </li>
 <li class="toctree-l1"><a class="reference internal" href="s100.html">s100py.s100 Module</a></li>
 <li class="toctree-l1"><a class="reference internal" href="s102.html">s100py.s102 Module</a></li>
+<li class="toctree-l1"><a class="reference internal" href="extending_the_api.html">Extending the API</a></li>
+<li class="toctree-l1"><a class="reference internal" href="sample_api.html">An Example API</a></li>
 </ul>
 
   <h4>Previous topic</h4>
-  <p class="topless"><a href="extending_the_api.html"
-                        title="previous chapter">Extending the API</a></p>
+  <p class="topless"><a href="s100py.html"
+                        title="previous chapter">s100py</a></p>
   <h4>Next topic</h4>
   <p class="topless"><a href="api/s100py.s1xx.S1XXFile.html"
                         title="next chapter">S1XXFile</a></p>
@@ -164,7 +165,7 @@ implementing one of the S100 family of specifications.</p>
           <a href="api/s100py.s1xx.S1XXFile.html" title="S1XXFile"
              >next</a> |</li>
         <li class="right" >
-          <a href="extending_the_api.html" title="Extending the API"
+          <a href="s100py.html" title="s100py"
              >previous</a> |</li>
         <li class="nav-item nav-item-0"><a href="s100py.html">Pydro  documentation</a> &#187;</li> 
       </ul>
