+++ s100py/docs/html/search.html
@@ -73,10 +73,11 @@
         <div class="sphinxsidebarwrapper">
   <h3><a href="s100py.html">Table of Contents</a></h3>
   <ul>
-<li class="toctree-l1"><a class="reference internal" href="extending_the_api.html">Extending the API</a></li>
 <li class="toctree-l1"><a class="reference internal" href="s1xx.html">s100py.s1xx Module</a></li>
 <li class="toctree-l1"><a class="reference internal" href="s100.html">s100py.s100 Module</a></li>
 <li class="toctree-l1"><a class="reference internal" href="s102.html">s100py.s102 Module</a></li>
+<li class="toctree-l1"><a class="reference internal" href="extending_the_api.html">Extending the API</a></li>
+<li class="toctree-l1"><a class="reference internal" href="sample_api.html">An Example API</a></li>
 </ul>
 
         </div>
