+++ s100py/docs/html/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({docnames:["api/s100py.s100.COMMON_POINT_RULE","api/s100py.s100.DATA_CODING_FORMAT","api/s100py.s100.DirectPosition","api/s100py.s100.FeatureInstance","api/s100py.s100.FeatureInstanceTimeSeries","api/s100py.s100.FeatureInstance_Format_2","api/s100py.s100.GeographicBoundingBox","api/s100py.s100.GeographicExtent","api/s100py.s100.GridCoordinate","api/s100py.s100.GridEnvelope","api/s100py.s100.INTERPOLATION_TYPE","api/s100py.s100.Point","api/s100py.s100.S100File","api/s100py.s100.S100Root","api/s100py.s100.S100_FeatureContainer","api/s100py.s100.SEQUENCING_RULE_TYPE","api/s100py.s100.SequenceRule","api/s100py.s100.VERTICAL_DATUM","api/s100py.s100.VertexPoint","api/s100py.s102.BathymetryContainer","api/s100py.s102.BathymetryCoverage","api/s100py.s102.BathymetryCoverage_dataset","api/s100py.s102.BathymetryCoverages_List","api/s100py.s102.BathymetryFeatureInstance","api/s100py.s102.BathymetryGroup_List","api/s100py.s102.BathymetryValueRecord","api/s100py.s102.BathymetryValues","api/s100py.s102.BathymetryValuesList","api/s100py.s102.DiscoveryMetadata","api/s100py.s102.FeatureCodes","api/s100py.s102.FeatureInformation","api/s100py.s102.FeatureInformation_dataset","api/s100py.s102.S102File","api/s100py.s102.S102Root","api/s100py.s102.S102_MetadataList_base","api/s100py.s102.SurfaceCorrectionValues","api/s100py.s102.TilingScheme","api/s100py.s102.TrackingListContainer","api/s100py.s102.TrackingListCoverage","api/s100py.s102.TrackingListCoverage_dataset","api/s100py.s102.TrackingListCoverages_List","api/s100py.s102.TrackingListGroup_List","api/s100py.s102.TrackingListSet_List","api/s100py.s102.TrackingListValues","api/s100py.s102.TrackingListValues_List","api/s100py.s1xx.S1XXFile","api/s100py.s1xx.S1XX_Attributes_base","api/s100py.s1xx.S1XX_Dataset_base","api/s100py.s1xx.S1XX_MetadataList_base","api/s100py.s1xx.S1XX_WritesOwnGroup_base","extending_the_api","s100","s100py","s102","s1xx"],envversion:{"sphinx.domains.c":1,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":1,"sphinx.domains.index":1,"sphinx.domains.javascript":1,"sphinx.domains.math":2,"sphinx.domains.python":1,"sphinx.domains.rst":1,"sphinx.domains.std":1,sphinx:56},filenames:["api\\s100py.s100.COMMON_POINT_RULE.rst","api\\s100py.s100.DATA_CODING_FORMAT.rst","api\\s100py.s100.DirectPosition.rst","api\\s100py.s100.FeatureInstance.rst","api\\s100py.s100.FeatureInstanceTimeSeries.rst","api\\s100py.s100.FeatureInstance_Format_2.rst","api\\s100py.s100.GeographicBoundingBox.rst","api\\s100py.s100.GeographicExtent.rst","api\\s100py.s100.GridCoordinate.rst","api\\s100py.s100.GridEnvelope.rst","api\\s100py.s100.INTERPOLATION_TYPE.rst","api\\s100py.s100.Point.rst","api\\s100py.s100.S100File.rst","api\\s100py.s100.S100Root.rst","api\\s100py.s100.S100_FeatureContainer.rst","api\\s100py.s100.SEQUENCING_RULE_TYPE.rst","api\\s100py.s100.SequenceRule.rst","api\\s100py.s100.VERTICAL_DATUM.rst","api\\s100py.s100.VertexPoint.rst","api\\s100py.s102.BathymetryContainer.rst","api\\s100py.s102.BathymetryCoverage.rst","api\\s100py.s102.BathymetryCoverage_dataset.rst","api\\s100py.s102.BathymetryCoverages_List.rst","api\\s100py.s102.BathymetryFeatureInstance.rst","api\\s100py.s102.BathymetryGroup_List.rst","api\\s100py.s102.BathymetryValueRecord.rst","api\\s100py.s102.BathymetryValues.rst","api\\s100py.s102.BathymetryValuesList.rst","api\\s100py.s102.DiscoveryMetadata.rst","api\\s100py.s102.FeatureCodes.rst","api\\s100py.s102.FeatureInformation.rst","api\\s100py.s102.FeatureInformation_dataset.rst","api\\s100py.s102.S102File.rst","api\\s100py.s102.S102Root.rst","api\\s100py.s102.S102_MetadataList_base.rst","api\\s100py.s102.SurfaceCorrectionValues.rst","api\\s100py.s102.TilingScheme.rst","api\\s100py.s102.TrackingListContainer.rst","api\\s100py.s102.TrackingListCoverage.rst","api\\s100py.s102.TrackingListCoverage_dataset.rst","api\\s100py.s102.TrackingListCoverages_List.rst","api\\s100py.s102.TrackingListGroup_List.rst","api\\s100py.s102.TrackingListSet_List.rst","api\\s100py.s102.TrackingListValues.rst","api\\s100py.s102.TrackingListValues_List.rst","api\\s100py.s1xx.S1XXFile.rst","api\\s100py.s1xx.S1XX_Attributes_base.rst","api\\s100py.s1xx.S1XX_Dataset_base.rst","api\\s100py.s1xx.S1XX_MetadataList_base.rst","api\\s100py.s1xx.S1XX_WritesOwnGroup_base.rst","extending_the_api.rst","s100.rst","s100py.rst","s102.rst","s1xx.rst"],objects:{"s100py.s100":{COMMON_POINT_RULE:[0,1,1,""],DATA_CODING_FORMAT:[1,1,1,""],DirectPosition:[2,1,1,""],FeatureInstance:[3,1,1,""],FeatureInstanceTimeSeries:[4,1,1,""],FeatureInstance_Format_2:[5,1,1,""],GeographicBoundingBox:[6,1,1,""],GeographicExtent:[7,1,1,""],GridCoordinate:[8,1,1,""],GridEnvelope:[9,1,1,""],INTERPOLATION_TYPE:[10,1,1,""],Point:[11,1,1,""],S100File:[12,1,1,""],S100Root:[13,1,1,""],S100_FeatureContainer:[14,1,1,""],SEQUENCING_RULE_TYPE:[15,1,1,""],SequenceRule:[16,1,1,""],VERTICAL_DATUM:[17,1,1,""],VertexPoint:[18,1,1,""]},"s100py.s100.COMMON_POINT_RULE":{all:[0,2,1,""],average:[0,2,1,""],high:[0,2,1,""],low:[0,2,1,""]},"s100py.s100.DirectPosition":{__init__:[2,3,1,""],coordinate:[2,2,1,""],dimension:[2,2,1,""]},"s100py.s100.FeatureInstance":{__init__:[3,3,1,""],instance_chunking:[3,2,1,""],num_grp:[3,2,1,""],vertical_extent_maximum_z:[3,2,1,""],vertical_extent_minimum_z:[3,2,1,""]},"s100py.s100.FeatureInstanceTimeSeries":{__init__:[4,3,1,""],date_time_of_first_record:[4,2,1,""],date_time_of_last_record:[4,2,1,""],number_of_times:[4,2,1,""],time_record_interval:[4,2,1,""]},"s100py.s100.FeatureInstance_Format_2":{__init__:[5,3,1,""],grid_origin_latitude:[5,2,1,""],grid_origin_longitude:[5,2,1,""],grid_origin_vertical:[5,2,1,""],grid_spacing_latitudinal:[5,2,1,""],grid_spacing_longitudinal:[5,2,1,""],grid_spacing_vertical:[5,2,1,""],num_points_latitudinal:[5,2,1,""],num_points_longitudinal:[5,2,1,""],num_points_vertical:[5,2,1,""],start_sequence:[5,2,1,""]},"s100py.s100.GeographicBoundingBox":{__init__:[6,3,1,""],east_bound_longitude:[6,2,1,""],north_bound_latitude:[6,2,1,""],south_bound_latitude:[6,2,1,""],west_bound_longitude:[6,2,1,""]},"s100py.s100.GeographicExtent":{__init__:[7,3,1,""],extent_type_code:[7,2,1,""]},"s100py.s100.GridCoordinate":{__init__:[8,3,1,""],coord_values:[8,2,1,""]},"s100py.s100.GridEnvelope":{__init__:[9,3,1,""],high:[9,2,1,""],low:[9,2,1,""]},"s100py.s100.INTERPOLATION_TYPE":{barycentric:[10,2,1,""],bicubic:[10,2,1,""],bilinear:[10,2,1,""],biquadratic:[10,2,1,""],cubic:[10,2,1,""],discrete:[10,2,1,""],linear:[10,2,1,""],lostarea:[10,2,1,""],nearestneighbor:[10,2,1,""],quadratic:[10,2,1,""]},"s100py.s100.Point":{__init__:[11,3,1,""],position:[11,2,1,""]},"s100py.s100.S100File":{PRODUCT_SPECIFICATION:[12,2,1,""],__init__:[12,3,1,""]},"s100py.s100.S100Root":{__init__:[13,3,1,""],epoch:[13,2,1,""],geographic_identifier:[13,2,1,""],horizontal_datum_reference:[13,2,1,""],horizontal_datum_value:[13,2,1,""],issue_date:[13,2,1,""],issue_time:[13,2,1,""],meta_features:[13,2,1,""],metadata:[13,2,1,""],product_specification:[13,2,1,""],vertical_datum:[13,2,1,""]},"s100py.s100.S100_FeatureContainer":{__init__:[14,3,1,""],axis_names:[14,2,1,""],common_point_rule:[14,2,1,""],data_coding_format:[14,2,1,""],dimension:[14,2,1,""],horizontal_position_uncertainty:[14,2,1,""],num_instances:[14,2,1,""],sequencing_rule_scan_direction:[14,2,1,""],time_uncertainty:[14,2,1,""],vertical_uncertainty:[14,2,1,""]},"s100py.s100.SEQUENCING_RULE_TYPE":{CantorDiagonal:[15,2,1,""],Hilbert:[15,2,1,""],Morton:[15,2,1,""],boustrophedonic:[15,2,1,""],linear:[15,2,1,""],spiral:[15,2,1,""]},"s100py.s100.SequenceRule":{__init__:[16,3,1,""],scan_direction:[16,2,1,""],type:[16,2,1,""]},"s100py.s100.VERTICAL_DATUM":{HAT:[17,2,1,""],LAT:[17,2,1,""],LW:[17,2,1,""],MHHW:[17,2,1,""],MHW:[17,2,1,""],MHWS:[17,2,1,""],MLLW:[17,2,1,""],MLW:[17,2,1,""],MLWS:[17,2,1,""],MSL:[17,2,1,""],approximateLowestAstronomicalTide:[17,2,1,""],approximateMeanLowWater:[17,2,1,""],approximateMeanLowWaterSprings:[17,2,1,""],approximateMeanLowerLowWater:[17,2,1,""],approximateMeanSeaLevel:[17,2,1,""],equinoctialSpringLowWater:[17,2,1,""],highWater:[17,2,1,""],highWaterSprings:[17,2,1,""],higherHighWaterLargeTide:[17,2,1,""],highestAstronomicalTide:[17,2,1,""],indianSpringLowWater:[17,2,1,""],internationalGreatLakesDatum1985:[17,2,1,""],localDatum:[17,2,1,""],lowWater:[17,2,1,""],lowWaterSprings:[17,2,1,""],lowerLowWaterLargeTide:[17,2,1,""],lowestAstronomicalTide:[17,2,1,""],lowestLowWater:[17,2,1,""],lowestLowWaterSprings:[17,2,1,""],meanHighWater:[17,2,1,""],meanHighWaterSprings:[17,2,1,""],meanHigherHighWater:[17,2,1,""],meanLowWater:[17,2,1,""],meanLowWaterSprings:[17,2,1,""],meanLowerLowWater:[17,2,1,""],meanLowerLowWaterSprings:[17,2,1,""],meanSeaLevel:[17,2,1,""],meanWaterLevel:[17,2,1,""],nearlyHighestHighWater:[17,2,1,""],nearlyLowestLowWater:[17,2,1,""]},"s100py.s100.VertexPoint":{__init__:[18,3,1,""],geometry:[18,2,1,""],value:[18,2,1,""]},"s100py.s102":{BathymetryContainer:[19,1,1,""],BathymetryCoverage:[20,1,1,""],BathymetryCoverage_dataset:[21,1,1,""],BathymetryCoverages_List:[22,1,1,""],BathymetryFeatureInstance:[23,1,1,""],BathymetryGroup_List:[24,1,1,""],BathymetryValueRecord:[25,1,1,""],BathymetryValues:[26,1,1,""],BathymetryValuesList:[27,1,1,""],DiscoveryMetadata:[28,1,1,""],FeatureCodes:[29,1,1,""],FeatureInformation:[30,1,1,""],FeatureInformation_dataset:[31,1,1,""],S102File:[32,1,1,""],S102Root:[33,1,1,""],S102_MetadataList_base:[34,1,1,""],SurfaceCorrectionValues:[35,1,1,""],TilingScheme:[36,1,1,""],TrackingListContainer:[37,1,1,""],TrackingListCoverage:[38,1,1,""],TrackingListCoverage_dataset:[39,1,1,""],TrackingListCoverages_List:[40,1,1,""],TrackingListGroup_List:[41,1,1,""],TrackingListSet_List:[42,1,1,""],TrackingListValues:[43,1,1,""],TrackingListValues_List:[44,1,1,""]},"s100py.s102.BathymetryContainer":{__init__:[19,3,1,""],bathymetry_coverage:[19,2,1,""]},"s100py.s102.BathymetryCoverage":{__init__:[20,3,1,""],axis_names:[20,2,1,""],dimension:[20,2,1,""],extent:[20,2,1,""],maximimum_display_scale:[20,2,1,""],maximum_depth:[20,2,1,""],maximum_uncertainty:[20,2,1,""],minimum_depth:[20,2,1,""],minimum_display_scale:[20,2,1,""],minimum_uncertainty:[20,2,1,""],offset_vectors:[20,2,1,""],origin:[20,2,1,""],sequencing_rule:[20,2,1,""],start_sequence:[20,2,1,""],values:[20,2,1,""],write_format_str:[20,2,1,""]},"s100py.s102.BathymetryCoverage_dataset":{__init__:[21,3,1,""],metadata_name:[21,2,1,""]},"s100py.s102.BathymetryCoverages_List":{__init__:[22,3,1,""],metadata_name:[22,2,1,""]},"s100py.s102.BathymetryFeatureInstance":{__init__:[23,3,1,""],bathymetry_group:[23,2,1,""]},"s100py.s102.BathymetryGroup_List":{__init__:[24,3,1,""],metadata_name:[24,2,1,""]},"s100py.s102.BathymetryValueRecord":{__init__:[25,3,1,""],depth:[25,2,1,""],uncertainty:[25,2,1,""]},"s100py.s102.BathymetryValues":{__init__:[26,3,1,""],depth:[26,2,1,""],get_write_order:[26,3,1,""],metadata_name:[26,2,1,""],read:[26,3,1,""],uncertainty:[26,2,1,""],write:[26,3,1,""]},"s100py.s102.BathymetryValuesList":{__init__:[27,3,1,""],metadata_name:[27,2,1,""]},"s100py.s102.DiscoveryMetadata":{__init__:[28,3,1,""]},"s100py.s102.FeatureCodes":{__init__:[29,3,1,""],bathymetry_coverage_dataset:[29,2,1,""],feature_code:[29,2,1,""],feature_name:[29,2,1,""],tracking_list_coverage:[29,2,1,""]},"s100py.s102.FeatureInformation":{__init__:[30,3,1,""],closure:[30,2,1,""],code:[30,2,1,""],datatype:[30,2,1,""],fill_value:[30,2,1,""],lower:[30,2,1,""],name:[30,2,1,""],unit_of_measure:[30,2,1,""],upper:[30,2,1,""]},"s100py.s102.FeatureInformation_dataset":{__init__:[31,3,1,""]},"s100py.s102.S102File":{PRODUCT_SPECIFICATION:[32,2,1,""],__init__:[32,3,1,""]},"s100py.s102.S102Root":{__init__:[33,3,1,""],bathymetry_coverage:[33,2,1,""],feature_information:[33,2,1,""],tracking_list_coverage:[33,2,1,""]},"s100py.s102.S102_MetadataList_base":{__init__:[34,3,1,""],write_format_str:[34,2,1,""]},"s100py.s102.SurfaceCorrectionValues":{__init__:[35,3,1,""]},"s100py.s102.TilingScheme":{__init__:[36,3,1,""]},"s100py.s102.TrackingListContainer":{__init__:[37,3,1,""],tracking_list_coverage:[37,2,1,""]},"s100py.s102.TrackingListCoverage":{__init__:[38,3,1,""],common_point_rule:[38,2,1,""],domain_extent:[38,2,1,""],set:[38,2,1,""],write_format_str:[38,2,1,""]},"s100py.s102.TrackingListCoverage_dataset":{__init__:[39,3,1,""],metadata_name:[39,2,1,""]},"s100py.s102.TrackingListCoverages_List":{__init__:[40,3,1,""],metadata_name:[40,2,1,""]},"s100py.s102.TrackingListGroup_List":{__init__:[41,3,1,""],metadata_name:[41,2,1,""]},"s100py.s102.TrackingListSet_List":{__init__:[42,3,1,""],metadata_name:[42,2,1,""]},"s100py.s102.TrackingListValues":{__init__:[43,3,1,""],list_series:[43,2,1,""],track_code:[43,2,1,""]},"s100py.s102.TrackingListValues_List":{__init__:[44,3,1,""],metadata_name:[44,2,1,""]},"s100py.s1xx":{S1XXFile:[45,1,1,""],S1XX_Attributes_base:[46,1,1,""],S1XX_Dataset_base:[47,1,1,""],S1XX_MetadataList_base:[48,1,1,""],S1XX_WritesOwnGroup_base:[49,1,1,""]},"s100py.s1xx.S1XXFile":{__init__:[45,3,1,""],create_empty_metadata:[45,3,1,""],depth_keys:[45,2,1,""],get_depth_dataset:[45,3,1,""],get_depths:[45,3,1,""],group_level_keys:[45,2,1,""],print_depth_attributes:[45,3,1,""],print_overview:[45,3,1,""],read:[45,3,1,""],second_level_keys:[45,2,1,""],show_keys:[45,3,1,""],top_level_keys:[45,2,1,""],tracking_list_group_level:[45,2,1,""],tracking_list_second_level:[45,2,1,""],tracking_list_top_level:[45,2,1,""],value_level_keys:[45,2,1,""],write:[45,3,1,""]},"s100py.s1xx.S1XX_Attributes_base":{S1XX_version:[46,2,1,""],__init__:[46,3,1,""],add_metadata:[46,3,1,""],get_all_keys:[46,3,1,""],get_metadata:[46,3,1,""],get_s1xx_attr:[46,3,1,""],get_standard_keys:[46,3,1,""],get_standard_list_properties:[46,3,1,""],get_standard_properties:[46,3,1,""],get_standard_properties_mapping:[46,3,1,""],get_write_order:[46,3,1,""],initialize_properties:[46,3,1,""],read:[46,3,1,""],read_hdf5_attributes:[46,3,1,""],set_enum_attribute:[46,3,1,""],set_s1xx_attr:[46,3,1,""],write:[46,3,1,""],write_as_xml:[46,3,1,""]},"s100py.s1xx.S1XX_Dataset_base":{__init__:[47,3,1,""],append_new_item:[47,3,1,""],metadata_name:[47,2,1,""],read:[47,3,1,""],write:[47,3,1,""],write_as_xml:[47,3,1,""]},"s100py.s1xx.S1XX_MetadataList_base":{__init__:[48,3,1,""],append_new_item:[48,3,1,""],metadata_name:[48,2,1,""],read:[48,3,1,""],read_re_pattern:[48,2,1,""],write:[48,3,1,""],write_as_xml:[48,3,1,""],write_format_str:[48,2,1,""]},"s100py.s1xx.S1XX_WritesOwnGroup_base":{__init__:[49,3,1,""]},s100py:{s100:[51,0,0,"-"],s102:[53,0,0,"-"],s1xx:[54,0,0,"-"]}},objnames:{"0":["py","module","Python module"],"1":["py","class","Python class"],"2":["py","attribute","Python attribute"],"3":["py","method","Python method"]},objtypes:{"0":"py:module","1":"py:class","2":"py:attribute","3":"py:method"},terms:{"":[7,12,16,32,52],"02d":38,"03d":[20,34,48],"10c":[4,6,13,19],"abstract":6,"boolean":7,"byte":[12,32,45],"case":[30,50],"class":[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,52],"default":[12,16,20,32,38,45,50],"enum":[0,1,10,15,17,51],"final":52,"float":[3,5,6,14,20,25,30],"function":[10,46,50,52],"import":[26,46,52],"int":[2,3,4,5,12,13,14,20,32,43],"long":17,"new":[12,32,45,50],"return":[2,3,4,5,6,7,8,9,11,13,14,16,17,18,19,20,21,22,23,24,25,26,27,29,30,33,37,38,39,40,41,42,43,44,46,47,48,50,52],"super":46,"true":[12,32,45,52],"while":[9,12,16,17,30,32,45],For:[12,16,32,45,46,52],That:7,The:[4,6,7,8,9,11,12,16,18,19,20,23,25,26,27,30,32,33,37,38,43,45,46,47,48,49,52],There:[7,18,20,25,52],These:[18,38,52],Using:52,__init__:[2,3,4,5,6,7,8,9,11,12,13,14,16,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49],__version__:46,_attribut:[50,52],_attribute_nam:[46,50],_creat:50,_hl:[45,51,53,54],_nnn:52,_type:50,abc:[46,51,53,54],about:[45,54],abov:[50,52],access:[47,52],accord:[12,32,45],accur:[2,3,4,5,6,7,8,9,11,13,14,16,18,19,20,21,22,23,24,25,26,27,28,29,30,31,33,34,35,36,37,38,39,40,41,42,43,44,46,47,48,49],act:[22,48],actual:30,add:[45,52],add_metadata:46,addit:[12,32,45],aforement:52,alias:17,all:[0,12,16,32,45,46,52],allow:[12,16,20,32,38,45,52],along:10,also:[6,20,38,52],alt:50,alwai:[12,32,45],amd:52,ani:[10,26,47,52],aovvaw07qesny5urachwio1e4alu:9,api:[52,54],appear:52,append_new_item:[47,48,52],appli:[10,49],applic:[12,32,45,50],approxim:[12,32,45],approximatelowestastronomicaltid:17,approximatemeanlowerlowwat:17,approximatemeanlowwat:17,approximatemeanlowwaterspr:17,approximatemeansealevel:17,arbitrari:[38,52],arc:10,area:[7,10],arg:[12,21,22,24,27,28,31,32,34,39,40,41,42,44,45,47,48],arrai:[45,46,48,50],assign:[10,18,25,38],associ:[10,20,46],at_download:9,attr:[45,50],attribut:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,29,30,32,33,34,37,38,39,40,41,42,43,44,45,46,47,48,50,52],attribute_nam:46,atttribut:48,autodetermin:46,averag:0,axi:[8,16],axis_nam:[14,20],axisnam:8,barycentr:10,base:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,51,53,54],basic:[11,30,52],bathi:52,bathmetrycoverag:52,bathy_01:52,bathy_list:52,bathymetr:38,bathymetri:[19,23,25,33,43,52],bathymetry_coverag:[19,33,46,52],bathymetry_coverage_attribute_nam:52,bathymetry_coverage_cr:52,bathymetry_coverage_dataset:29,bathymetry_coverage_typ:52,bathymetry_group:23,bathymetrycontain:[52,53],bathymetrycoverag:[6,7,19,22,23,27,30,45,46,52,53],bathymetrycoverage_01:45,bathymetrycoverage_dataset:[29,52,53],bathymetrycoverages_list:[19,23,52,53],bathymetryfeatureinst:[52,53],bathymetrygroup:23,bathymetrygroup_list:[52,53],bathymetryvalu:[20,52,53],bathymetryvaluerecord:[52,53],bathymetryvalueslist:[52,53],becaus:52,been:[12,32,45,52],befor:[12,32,45],begin:52,being:52,believ:29,below:50,between:[10,12,32,45],bicub:10,bilinear:10,biquadrat:10,blank:[50,52],block:[12,32,45],bool:7,bottom:50,bound:[6,7,12,32,45],boundari:[6,7],boustrophedon:15,box:[6,7],built:46,button:50,cach:[12,32,45],call:[46,49,52],camel:[30,50],can:[9,12,17,32,45,49,52],cantordiagon:15,capit:16,carri:7,cell:[10,16,20,27],chang:50,charact:[14,20],characterstr:[16,43],child:[19,52],children:52,chunk:[12,32,45],click:50,closedinterv:30,closur:30,code:[10,29,30,50,52,54],codelist:16,collect:[51,53,54],collis:[12,32,45],column:9,combin:49,commetn:[],common_point_rul:[14,38,51,52],commonpointrul:38,commonstateobject:[51,53,54],complex:52,compon:7,composit:27,comput:10,conform:[16,20],connect:10,consist:11,contain:[19,22,23,33,37,43,45,50,51,52,53,54],coord_valu:8,coordin:[2,8,9,38],coordvalu:8,core:[12,32,45],corner:20,correct:29,correspond:[7,18,43],could:47,cover:7,coverag:[10,18,33,38,43,52],coveragesattribut:33,creat:[12,17,26,32,45,46,47,49,50,52,54],create_dataset:45,create_empty_metadata:45,create_group:45,creation:[12,32,45,50],ctrl:50,cubic:10,current:[12,32,33,45],cursor:50,curv:10,cv_discretepointcoverag:38,cv_geometryvaluepair:38,cv_gridcoordin:20,cv_gridenvelop:[6,7],cv_interpolationmethod:10,cv_sequencerul:[16,20],data:[7,9,11,12,16,18,19,20,23,25,27,30,32,37,45,47,48,49,50,52],data_coding_format:[14,51,52],datanam:52,dataname_attribute_nam:52,dataname_cr:52,dataname_typ:52,dataset:[2,8,12,14,16,18,20,24,26,29,30,32,45,46,47,49,50,52],datatyp:30,date_time_of_first_record:4,date_time_of_last_record:4,datum:17,def:50,defin:[6,12,25,32,38,43,45,50],definit:25,denot:30,depend:[12,32,45],depth:[18,24,25,26,27,30,38,43,45,47],depth_kei:45,depthcorrectiontyp:25,deriv:[18,49,52],describ:[7,9,10,20,30,43],descript:50,desir:[12,32,45],detail:[12,32,45,54],determin:[12,20,32,33,45,52],diagram:52,dict:46,dictionari:[45,46],differ:[12,32,45,52],dimens:[2,8,14,20],direct:[8,16,20],directli:54,directposit:[11,20,51,52],discoverymetadata:[52,53],discret:[10,18],disk:[12,32,45],display_nod:45,distanc:10,doc:[9,50],document:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,29,30,32,33,34,37,38,39,40,41,42,43,44,45,46,47,48],doe:50,domain:10,domain_ext:38,dot:52,driver:[12,32,45],due:[12,32,45],duplic:52,dure:[26,47],each:[8,18,20,24,27,38,52],earliest:[12,32,45],east:16,east_bound_longitud:[6,52],eastboundlongitud:52,edit:[38,50],editor:50,either:49,element:[6,7,8,20,30,43,46],elev:45,empti:[12,32,45,50,52],encapsul:52,encod:52,encompass:7,end:[26,47],entri:30,enum_typ:46,enumer:[0,1,10,15,16],envelop:9,epoch:[13,52],equinoctialspringlowwat:17,error:29,establish:[6,7],etc:[17,22,30,46,48],etree_object:[46,47,48],eventu:46,everyth:52,ex_ext:7,ex_geographicboundingbox:6,ex_geographicext:[6,7],exampl:5,exist:[12,32,45,50],expect:52,explan:[12,32,45],express:50,extend:[52,54],extending_the_api:[],extens:[10,52],extent:[7,9,20],extent_type_cod:[7,52],extenttypecod:[7,52],extra:[26,47],fail:[12,32,45],fals:[2,3,4,5,6,7,8,9,11,13,14,16,18,19,20,23,25,26,29,30,33,35,36,37,38,43,46,49,52],famili:[52,54],feat:47,featur:[4,10,16,30,33],feature_cod:29,feature_inform:[33,46],feature_information_cr:52,feature_nam:29,featurecod:[33,52,53],featureinform:[47,52,53],featureinformation_dataset:[21,39,52,53],featureinst:[4,5,51,52,53],featureinstance_format_2:[23,51,52,53],featureinstancetimeseri:[51,52],featurenam:29,fgdc:9,field:43,figur:[11,18,20,22,25,27,33,38,40],file:[9,12,17,26,32,45,46,47,48,50,51,52,53,54],fill:50,fill_empti:[2,3,4,5,6,7,8,9,11,13,14,16,18,19,20,23,25,26,29,30,33,35,36,37,38,43,46,49],fill_valu:30,first:[20,52],fit:[12,32,45],flush:[12,32,45],format:[5,52],frameword:9,framework:50,from:[4,6,7,8,11,12,13,16,17,18,19,26,27,32,33,38,45,46,47,48,50,52],fulli:[12,32,45],further:[16,54],gener:[46,52],geographic_identifi:13,geographicboundingbox:[3,13,51,52,53],geographicext:[6,51,52,53],geometr:[10,11],geometri:[18,38],get:[46,52],get_all_kei:46,get_config:[12,32,45],get_depth:45,get_depth_dataset:45,get_metadata:46,get_s1xx_attr:46,get_standard_kei:46,get_standard_list_properti:46,get_standard_properti:[46,52],get_standard_properties_map:[46,52],get_write_ord:[26,46],gi_frameworkdatastandard_part3_elev:9,give:50,global:[12,32,45],gm_point:11,going:52,gotten:9,gov:9,grab:52,grid:[5,8,9,10,16,18,20,25,27],grid_origin_latitud:5,grid_origin_longitud:5,grid_origin_vert:5,grid_spacing_latitudin:5,grid_spacing_longitudin:5,grid_spacing_vert:5,gridcoordin:[20,51,52],gridenvelop:[20,51,52],gridoriginlatitud:52,gridoriginlongitud:52,group:[4,12,19,22,24,26,32,33,45,46,47,48,49,50,51,52,53,54],group_001:45,group_f:46,group_level_kei:45,group_nnn:50,group_object:[26,46,47,48],groupf:30,guid:[12,32,45],h5py:[12,26,32,45,46,47,48,51,52,53,54],h5t_native_float:30,handl:52,has:[8,16,18,19,20,24,25,43,46,50,52],hash:[12,32,45],hat:17,have:[12,30,32,45,46,49,50,52],hdf5:[12,18,19,26,32,45,46,47,48,50,52],held:[9,24,52],help:[2,3,4,5,6,7,8,9,11,13,14,16,18,19,20,21,22,23,24,25,26,27,28,29,30,31,33,34,35,36,37,38,39,40,41,42,43,44,46,47,48,49],here:[22,50],high:[0,9],higherhighwaterlargetid:17,highestastronomicaltid:17,highlight:50,highwat:17,highwaterspr:17,hilbert:15,hit:50,hlobject:[51,53,54],hold:[8,18,19],holder:52,horizont:38,horizontal_datum_refer:13,horizontal_datum_valu:13,horizontal_position_uncertainti:14,how:[12,20,32,45,50],http:9,hyperlink:50,ideal:[12,32,45],identifi:[20,38],iho:[12,32],implement:[46,54],inclus:[12,32,45],increas:[12,32,45],indent:[26,45,46,47,48],index:[43,47],indianspringlowwat:17,indic:[6,7,9,12,16,18,32,45],inform:[7,33],inherit:[27,52],initi:[2,3,4,5,6,7,8,9,11,13,14,16,18,19,20,21,22,23,24,25,26,27,28,29,30,31,33,34,35,36,37,38,39,40,41,42,43,44,46,47,48,49,52],initialize_properti:[46,52],insert:45,insid:52,instanc:[4,47],instance_chunk:3,instead:[9,29],integ:[8,9,43],internationalgreatlakesdatum1985:17,interpol:10,interpolation_typ:[51,52],iso:[6,7,9,10,11,16,18,20,38],issue_d:13,issue_tim:13,item:[18,25,26,46,52],iter:[51,53,54],its:[49,52],itself:52,kei:[26,46],keyerror:52,keyword:[12,32,45],known:[10,16],kwarg:28,kywrd:[12,32,45],last:52,lat:17,latest:[12,32,45],latitid:52,latitud:[16,52],layer:16,least:[12,32,45],left:[9,20],legal:[12,32,45],length:10,let:52,level:[30,52],librari:[12,32,45],libver:[12,32,45],like:[12,22,29,32,45,46,48],line:[10,50,52],lineag:43,linear:[10,15,16,20],list:[6,7,16,18,19,22,23,24,26,30,33,37,38,43,46,47,48,50,52],list_seri:43,listseri:[18,43],liter:10,littl:52,live:50,livetempl:50,localdatum:17,locat:18,logic:49,longitud:16,lost:10,lostarea:10,low:[0,9],lower:[9,12,20,30,32,45],lowercas:16,lowerlowwaterlargetid:17,lowestastronomicaltid:17,lowestlowwat:17,lowestlowwaterspr:17,lowwat:17,lowwaterspr:17,lru:[12,32,45],mai:[7,43],make:[45,48,49,50],manag:52,map:[16,51,53,54],mappinghdf5:[51,53,54],match:[18,30,33],matrix:9,maximimum_display_scal:20,maximum:[12,32,45],maximum_depth:20,maximum_uncertainti:20,mean:[12,32,45],meanhigherhighwat:17,meanhighwat:17,meanhighwaterspr:17,meanlowerlowwat:17,meanlowerlowwaterspr:17,meanlowwat:17,meanlowwaterspr:17,meansealevel:17,meanwaterlevel:17,measur:30,member:[50,52],memori:[12,32,45],meta_featur:13,metaclass:7,metadata:[6,7,13,30,38,43,45,52],metadata_nam:[21,22,24,26,27,39,40,41,42,44,47,48],method:[10,26,45,46,47,48,52],metr:30,mhhw:17,mhw:17,miniconda36:52,minimum_depth:20,minimum_display_scal:20,minimum_uncertainti:20,miss:30,mllw:17,mlw:17,mode:[12,32,45],model:38,modifi:[50,54],modul:52,more:[38,52],morton:[15,16,20],most:[29,52],mpio:[12,32,45],msl:17,multipl:[46,49,50,52],must:[12,32,45,46],mutablemap:[51,53,54],mutablemappinghdf5:[51,53,54],name:[12,16,17,19,26,29,30,32,33,45,46,47,48,49,50,52],name_nnn:52,navo:29,ndarrai:[2,8,14,16,18,20,26,29],nearest:10,nearestneighbor:10,nearlyhighesthighwat:17,nearlylowestlowwat:17,necessari:38,need:49,next:52,nnn:[22,24,48],noaa:52,nomenclatur:52,non:[12,32,45],none:[12,26,32,45,46],north:[16,52],north_bound_latitud:[6,52],northboundlatitud:52,notat:52,note:[12,17,20,32,45,48,52],nov:46,now:52,num_grp:3,num_inst:14,num_points_latitudin:5,num_points_longitudin:5,num_points_vert:5,number:[8,12,32,43,45,52],number_of_tim:4,numpi:[46,50],obj:45,object:[10,11,12,17,19,23,26,32,37,45,46,47,52],occur:46,occurr:[50,52],off:19,offici:17,offset:[8,9,20],offset_vector:20,offsetvector:[9,20],often:[12,32,45],omit:52,onc:[12,32,45],one:[7,8,11,38,50,52,54],onli:[7,10,11,12,16,20,32,45,50,54],open:[12,32,45],opt:[21,22,24,27,31,34,39,40,41,42,44,47,48],option:[7,12,16,20,32,38,43,45,52],order:[8,12,16,20,26,27,32,45,46,47],ordereddict:52,organ:38,origin:[8,9,20],other:[12,16,20,22,32,45],otherwis:[12,26,29,32,45,46],our:52,out:[26,47],overrid:[26,43,46],overridden:[18,46],overwrit:52,own:[49,52],packag:52,pad:52,pair:[10,38],parallel:30,paramet:[10,26,46,47,48],parent:49,part:48,pass:[12,32,45],past:50,penal:[12,32,45],pep8:52,per:[12,32,45],perform:[12,32,45],pertain:52,place:[50,52],plain:30,plu:50,point:[10,16,18,20,25,38,51,52],polici:[12,32,45],polygon:[7,10],popul:52,portion:7,posit:[10,11,16],possibl:19,pre:49,predict:49,preempt:[12,32,45],preemption:[12,32,45],prescrib:52,present:7,press:50,primari:[45,52],prime:[12,32,45],print:52,print_depth_attribut:45,print_overview:45,product:52,product_specif:[12,13,32],project:9,properti:[46,50,52],provid:27,publish:46,put:[50,52],pycharm:50,pydrotrunk:52,python3:52,python:[48,50,52],quad:[16,20],quadrat:10,quadrilater:10,raw:[12,32,45],rdcc_nbyte:[12,32,45],rdcc_nslot:[12,32,45],rdcc_w0:[12,32,45],read:[9,12,26,32,45,46,47,48,50,52],read_hdf5_attribut:46,read_re_pattern:48,readonli:[12,32,45],real:9,realli:[48,52],reason:43,recent:52,recommend:[12,32,45],record:[16,18,20,25,38,48],recurs:52,reduc:[12,32,45],refactor:52,refer:52,referenc:[18,52],regular:[16,20],relat:[27,52],relationship:27,remov:29,repreent:48,repres:[38,48],requir:[12,20,32,45],right:[9,19,50],root:[12,19,32,33,45,46,52],row:9,rtype:[19,23,26,30,37,46],rule:[12,32,45],s100:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,23,33,35,37,38,48,50,52,53,54],s100_cv_interpolationmethod:10,s100_featurecontain:[19,37,51,52,53],s100_grid:27,s100_point:38,s100_pointcoverag:38,s100file:[51,52],s100py:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49],s100root:[33,51,52,53],s102:[5,16,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,46,47,48,50,52],s102_bathymetrycoverag:45,s102_bathymetrycoverage_01:45,s102_bathymetryvalu:[25,27],s102_bathymetryvaluerecord:25,s102_dataidentif:25,s102_elev:45,s102_grid:45,s102_grid_01:45,s102_metadatalist_bas:[19,22,23,24,27,37,38,40,41,42,44,52,53],s102_trackinglistvalu:18,s102file:[52,53],s102root:[52,53],s1xx:[2,7,8,9,11,12,14,16,18,20,25,26,28,29,30,31,32,34,36,38,45,46,47,48,49,51,52,53],s1xx_attributes_bas:[2,7,8,9,11,14,16,18,20,25,28,29,30,33,36,38,49,50,51,52,53,54],s1xx_dataset_bas:[31,46,50,52,53,54],s1xx_metadatalist_bas:[34,46,50,52,53,54],s1xx_name:46,s1xx_version:46,s1xx_writesowngroup_bas:[26,47,48,52,53,54],s1xxfile:[12,32,51,52,53,54],safe:[12,32,45],sai:[50,52],same:[8,9,12,32,45],sampl:52,scan:16,scan_direct:16,scandirect:16,scheme:50,sec2:[12,32,45],sec:[9,38],second_level_kei:45,sect:29,section:20,see:[2,3,4,5,6,7,8,9,11,12,13,14,16,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,52,54],seem:9,segment:10,select:[12,32,45,50],self:[2,3,4,5,6,7,8,9,11,13,14,16,18,19,20,21,22,23,24,25,26,27,28,29,30,31,33,34,35,36,37,38,39,40,41,42,43,44,46,47,48,49,50,52],sequenc:[8,14,16,18,20,25,27],sequencerul:[20,51,52],sequencetyp:16,sequencing_rul:20,sequencing_rule_scan_direct:14,sequencing_rule_typ:[51,52],sequencingrul:20,set:[7,12,16,22,32,38,45,46,50,52],set_enum_attribut:46,set_s1xx_attr:46,setter:50,shall:[8,16,18,20,25],shorthand:17,should:[12,18,29,32,45,47,52,54],show:52,show_kei:45,side:50,signatur:[2,3,4,5,6,7,8,9,11,13,14,16,18,19,20,21,22,23,24,25,26,27,28,29,30,31,33,34,35,36,37,38,39,40,41,42,43,44,46,47,48,49],simpl:[16,20],sinc:9,singl:8,site:52,size:[9,12,16,20,32,45,51,53,54],skeleton:52,slightli:[12,32,45],slot:[12,32,45],snakecas:50,softwar:29,some:[47,52],sound:38,south:16,south_bound_latitud:6,space:20,spatial:[6,7,38],spec:[4,13,19,47,48,52],special:49,specif:[8,52,54],specifi:[9,16,20,26,47,49,50],speed:50,spiral:15,standard:[9,16,20],start_sequ:[5,20],startsequ:20,stdio:[12,32,45],still:[12,32,45],store:[30,46,47],str:[3,4,5,13,14,16,21,22,24,26,27,30,38,39,40,41,42,43,44,47,48,50],strategi:[12,32,45],strictli:[12,32,45],string:[12,14,16,17,20,30,32,43,45],structur:52,sub:[46,49,50],subgroup:46,subtyp:[6,7],summari:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,29,30,32,33,34,37,38,39,40,41,42,43,44,45,46,47,48],suppli:48,surfacecorrectionvalu:[43,52,53],swmr:[12,32,45],tabl:[4,6,13,19,20,22,29,30,36,37,40],take:[48,50],taken:11,tell:50,temp:52,templat:50,test:52,text:[30,43,50],than:[12,32,45,52],thei:[22,52,54],them:[47,52],therefor:38,therein:52,thi:[9,12,16,19,20,22,24,26,29,30,32,43,45,46,47,48,50,52,54],thiessen:10,thing:[46,47,49],think:9,three:52,thumb:[12,32,45],tilingschem:[52,53],time:[12,32,45],time_record_interv:4,time_uncertainti:14,tin:10,too:[26,46,47,48],top:[30,52],top_level_kei:45,total:[12,32,45],traceback:52,track:[6,7,12,18,32,33,37,45],track_cod:43,track_ord:[12,32,45],trackcod:43,tracking_list_coverag:[29,33,37,46],tracking_list_group_level:45,tracking_list_second_level:45,tracking_list_top_level:45,trackinglistcontain:[52,53],trackinglistcoverag:[30,37,45,46,52,53],trackinglistcoverage_dataset:[29,52,53],trackinglistcoverages_list:[37,52,53],trackinglistgroup_list:[52,53],trackinglistset_list:[52,53],trackinglistvalu:[38,52,53],trackinglistvalues_list:[38,52,53],trail:52,treat:[12,32,45],tree:[16,20],truncat:[12,32,45],tupl:46,two:[10,18,25,33,52],type:[2,3,4,5,6,7,8,9,10,11,13,14,16,18,19,20,21,22,23,24,25,26,27,28,29,30,31,33,34,35,36,37,38,39,40,41,42,43,44,46,47,48,49,50,52],uml:48,uncertainti:[18,24,25,26,30,43,47],under:[12,30,32,45,50],underscor:52,union:[2,8,14,16,18,20,26,29],unit:30,unit_of_measur:30,unnecessari:9,unord:[26,47],upper:[9,30],use:[7,12,32,45,46,50],used:[5,6,7,12,16,20,30,32,45,47,52,54],user:[12,32,43,45],userblock:[12,32,45],uses:52,usg:9,using:[9,10,26,47,48,50,52,54],val:[46,50],valu:[7,8,9,10,12,16,18,20,25,26,27,30,32,38,43,45,46,47,50,52],value_level_kei:45,variabl:50,vector:20,version:[12,32,45,46],vertexpoint:[35,51,52,53],vertic:17,vertical_datum:[13,51,52],vertical_extent_maximum_z:3,vertical_extent_minimum_z:3,vertical_uncertainti:14,verticaluncertaintytyp:25,via:[19,23,37],wai:52,want:48,weight:[12,32,45],well:[22,38],were:18,west:16,west_bound_longitud:[6,52],westboundlongitud:46,what:[50,52],when:[7,12,32,45,52],where:[7,52],whether:7,which:[9,12,16,18,19,25,27,32,33,43,45,46,48,52],within:[6,7,10,16,30],without:49,word:50,work:46,world:9,would:[9,47,52],write:[12,26,32,45,46,47,48,50],write_as_xml:[46,47,48],write_format_str:[20,34,38,48],writer:49,written:[12,17,32,45,46,52],www:9,xml:46,yes:52,you:[12,32,45,52],your:[12,32,45,50],zero:[50,52]},titles:["COMMON_POINT_RULE","DATA_CODING_FORMAT","DirectPosition","FeatureInstance","FeatureInstanceTimeSeries","FeatureInstance_Format_2","GeographicBoundingBox","GeographicExtent","GridCoordinate","GridEnvelope","INTERPOLATION_TYPE","Point","S100File","S100Root","S100_FeatureContainer","SEQUENCING_RULE_TYPE","SequenceRule","VERTICAL_DATUM","VertexPoint","BathymetryContainer","BathymetryCoverage","BathymetryCoverage_dataset","BathymetryCoverages_List","BathymetryFeatureInstance","BathymetryGroup_List","BathymetryValueRecord","BathymetryValues","BathymetryValuesList","DiscoveryMetadata","FeatureCodes","FeatureInformation","FeatureInformation_dataset","S102File","S102Root","S102_MetadataList_base","SurfaceCorrectionValues","TilingScheme","TrackingListContainer","TrackingListCoverage","TrackingListCoverage_dataset","TrackingListCoverages_List","TrackingListGroup_List","TrackingListSet_List","TrackingListValues","TrackingListValues_List","S1XXFile","S1XX_Attributes_base","S1XX_Dataset_base","S1XX_MetadataList_base","S1XX_WritesOwnGroup_base","Extending the API","s100py.s100 Module","s100py","s100py.s102 Module","s100py.s1xx Module"],titleterms:{"class":[51,53,54],api:50,bathymetrycontain:19,bathymetrycoverag:20,bathymetrycoverage_dataset:21,bathymetrycoverages_list:22,bathymetryfeatureinst:23,bathymetrygroup_list:24,bathymetryvalu:26,bathymetryvaluerecord:25,bathymetryvalueslist:27,common_point_rul:0,data_coding_format:1,diagram:[51,53,54],directposit:2,discoverymetadata:28,extend:50,featurecod:29,featureinform:30,featureinformation_dataset:31,featureinst:3,featureinstance_format_2:5,featureinstancetimeseri:4,geographicboundingbox:6,geographicext:7,gridcoordin:8,gridenvelop:9,inherit:[51,53,54],interpolation_typ:10,modul:[51,53,54],point:11,s100:51,s100_featurecontain:14,s100file:12,s100py:[51,52,53,54],s100root:13,s102:53,s102_metadatalist_bas:34,s102file:32,s102root:33,s1xx:54,s1xx_attributes_bas:46,s1xx_dataset_bas:47,s1xx_metadatalist_bas:48,s1xx_writesowngroup_bas:49,s1xxfile:45,sequencerul:16,sequencing_rule_typ:15,surfacecorrectionvalu:35,tilingschem:36,trackinglistcontain:37,trackinglistcoverag:38,trackinglistcoverage_dataset:39,trackinglistcoverages_list:40,trackinglistgroup_list:41,trackinglistset_list:42,trackinglistvalu:43,trackinglistvalues_list:44,vertexpoint:18,vertical_datum:17}})
+Search.setIndex({docnames:["api/s100py.s100.COMMON_POINT_RULE","api/s100py.s100.DATA_CODING_FORMAT","api/s100py.s100.DirectPosition","api/s100py.s100.FeatureInformation","api/s100py.s100.FeatureInstance","api/s100py.s100.FeatureInstanceTimeSeries","api/s100py.s100.FeatureInstance_Format_2","api/s100py.s100.GeographicBoundingBox","api/s100py.s100.GeographicExtent","api/s100py.s100.GridCoordinate","api/s100py.s100.GridEnvelope","api/s100py.s100.INTERPOLATION_TYPE","api/s100py.s100.Point","api/s100py.s100.S100File","api/s100py.s100.S100Root","api/s100py.s100.S100_FeatureContainer","api/s100py.s100.SEQUENCING_RULE_TYPE","api/s100py.s100.SequenceRule","api/s100py.s100.VERTICAL_DATUM","api/s100py.s100.VertexPoint","api/s100py.s102.BathymetryContainer","api/s100py.s102.BathymetryCoverage","api/s100py.s102.BathymetryCoverage_dataset","api/s100py.s102.BathymetryCoverages_List","api/s100py.s102.BathymetryFeatureInstance","api/s100py.s102.BathymetryGroup_List","api/s100py.s102.BathymetryValues","api/s100py.s102.DiscoveryMetadata","api/s100py.s102.FeatureCodes","api/s100py.s102.FeatureInformation_dataset","api/s100py.s102.S102File","api/s100py.s102.S102Root","api/s100py.s102.S102_MetadataList_base","api/s100py.s102.SurfaceCorrectionValues","api/s100py.s102.TilingScheme","api/s100py.s102.TrackingListContainer","api/s100py.s102.TrackingListCoverage","api/s100py.s102.TrackingListCoverage_dataset","api/s100py.s102.TrackingListCoverages_List","api/s100py.s102.TrackingListGroup_List","api/s100py.s102.TrackingListSet_List","api/s100py.s102.TrackingListValues","api/s100py.s102.TrackingListValues_List","api/s100py.s1xx.S1XXFile","api/s100py.s1xx.S1XX_Attributes_base","api/s100py.s1xx.S1XX_Dataset_base","api/s100py.s1xx.S1XX_MetadataList_base","api/s100py.s1xx.S1XX_WritesOwnGroup_base","extending_the_api","s100","s100py","s102","s1xx","sample_api"],envversion:{"sphinx.domains.c":1,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":1,"sphinx.domains.index":1,"sphinx.domains.javascript":1,"sphinx.domains.math":2,"sphinx.domains.python":1,"sphinx.domains.rst":1,"sphinx.domains.std":1,sphinx:56},filenames:["api\\s100py.s100.COMMON_POINT_RULE.rst","api\\s100py.s100.DATA_CODING_FORMAT.rst","api\\s100py.s100.DirectPosition.rst","api\\s100py.s100.FeatureInformation.rst","api\\s100py.s100.FeatureInstance.rst","api\\s100py.s100.FeatureInstanceTimeSeries.rst","api\\s100py.s100.FeatureInstance_Format_2.rst","api\\s100py.s100.GeographicBoundingBox.rst","api\\s100py.s100.GeographicExtent.rst","api\\s100py.s100.GridCoordinate.rst","api\\s100py.s100.GridEnvelope.rst","api\\s100py.s100.INTERPOLATION_TYPE.rst","api\\s100py.s100.Point.rst","api\\s100py.s100.S100File.rst","api\\s100py.s100.S100Root.rst","api\\s100py.s100.S100_FeatureContainer.rst","api\\s100py.s100.SEQUENCING_RULE_TYPE.rst","api\\s100py.s100.SequenceRule.rst","api\\s100py.s100.VERTICAL_DATUM.rst","api\\s100py.s100.VertexPoint.rst","api\\s100py.s102.BathymetryContainer.rst","api\\s100py.s102.BathymetryCoverage.rst","api\\s100py.s102.BathymetryCoverage_dataset.rst","api\\s100py.s102.BathymetryCoverages_List.rst","api\\s100py.s102.BathymetryFeatureInstance.rst","api\\s100py.s102.BathymetryGroup_List.rst","api\\s100py.s102.BathymetryValues.rst","api\\s100py.s102.DiscoveryMetadata.rst","api\\s100py.s102.FeatureCodes.rst","api\\s100py.s102.FeatureInformation_dataset.rst","api\\s100py.s102.S102File.rst","api\\s100py.s102.S102Root.rst","api\\s100py.s102.S102_MetadataList_base.rst","api\\s100py.s102.SurfaceCorrectionValues.rst","api\\s100py.s102.TilingScheme.rst","api\\s100py.s102.TrackingListContainer.rst","api\\s100py.s102.TrackingListCoverage.rst","api\\s100py.s102.TrackingListCoverage_dataset.rst","api\\s100py.s102.TrackingListCoverages_List.rst","api\\s100py.s102.TrackingListGroup_List.rst","api\\s100py.s102.TrackingListSet_List.rst","api\\s100py.s102.TrackingListValues.rst","api\\s100py.s102.TrackingListValues_List.rst","api\\s100py.s1xx.S1XXFile.rst","api\\s100py.s1xx.S1XX_Attributes_base.rst","api\\s100py.s1xx.S1XX_Dataset_base.rst","api\\s100py.s1xx.S1XX_MetadataList_base.rst","api\\s100py.s1xx.S1XX_WritesOwnGroup_base.rst","extending_the_api.rst","s100.rst","s100py.rst","s102.rst","s1xx.rst","sample_api.rst"],objects:{"s100py.s100":{COMMON_POINT_RULE:[0,1,1,""],DATA_CODING_FORMAT:[1,1,1,""],DirectPosition:[2,1,1,""],FeatureInformation:[3,1,1,""],FeatureInstance:[4,1,1,""],FeatureInstanceTimeSeries:[5,1,1,""],FeatureInstance_Format_2:[6,1,1,""],GeographicBoundingBox:[7,1,1,""],GeographicExtent:[8,1,1,""],GridCoordinate:[9,1,1,""],GridEnvelope:[10,1,1,""],INTERPOLATION_TYPE:[11,1,1,""],Point:[12,1,1,""],S100File:[13,1,1,""],S100Root:[14,1,1,""],S100_FeatureContainer:[15,1,1,""],SEQUENCING_RULE_TYPE:[16,1,1,""],SequenceRule:[17,1,1,""],VERTICAL_DATUM:[18,1,1,""],VertexPoint:[19,1,1,""]},"s100py.s100.COMMON_POINT_RULE":{all:[0,2,1,""],average:[0,2,1,""],high:[0,2,1,""],low:[0,2,1,""]},"s100py.s100.DirectPosition":{__init__:[2,3,1,""],coordinate:[2,2,1,""],dimension:[2,2,1,""]},"s100py.s100.FeatureInformation":{__init__:[3,3,1,""],closure:[3,2,1,""],code:[3,2,1,""],datatype:[3,2,1,""],fill_value:[3,2,1,""],lower:[3,2,1,""],name:[3,2,1,""],unit_of_measure:[3,2,1,""],upper:[3,2,1,""]},"s100py.s100.FeatureInstance":{__init__:[4,3,1,""],instance_chunking:[4,2,1,""],num_grp:[4,2,1,""],vertical_extent_maximum_z:[4,2,1,""],vertical_extent_minimum_z:[4,2,1,""]},"s100py.s100.FeatureInstanceTimeSeries":{__init__:[5,3,1,""],date_time_of_first_record:[5,2,1,""],date_time_of_last_record:[5,2,1,""],number_of_times:[5,2,1,""],time_record_interval:[5,2,1,""]},"s100py.s100.FeatureInstance_Format_2":{__init__:[6,3,1,""],grid_origin_latitude:[6,2,1,""],grid_origin_longitude:[6,2,1,""],grid_origin_vertical:[6,2,1,""],grid_spacing_latitudinal:[6,2,1,""],grid_spacing_longitudinal:[6,2,1,""],grid_spacing_vertical:[6,2,1,""],num_points_latitudinal:[6,2,1,""],num_points_longitudinal:[6,2,1,""],num_points_vertical:[6,2,1,""],start_sequence:[6,2,1,""]},"s100py.s100.GeographicBoundingBox":{__init__:[7,3,1,""],east_bound_longitude:[7,2,1,""],north_bound_latitude:[7,2,1,""],south_bound_latitude:[7,2,1,""],west_bound_longitude:[7,2,1,""]},"s100py.s100.GeographicExtent":{__init__:[8,3,1,""],extent_type_code:[8,2,1,""]},"s100py.s100.GridCoordinate":{__init__:[9,3,1,""],coord_values:[9,2,1,""]},"s100py.s100.GridEnvelope":{__init__:[10,3,1,""],high:[10,2,1,""],low:[10,2,1,""]},"s100py.s100.INTERPOLATION_TYPE":{barycentric:[11,2,1,""],bicubic:[11,2,1,""],bilinear:[11,2,1,""],biquadratic:[11,2,1,""],cubic:[11,2,1,""],discrete:[11,2,1,""],linear:[11,2,1,""],lostarea:[11,2,1,""],nearestneighbor:[11,2,1,""],quadratic:[11,2,1,""]},"s100py.s100.Point":{__init__:[12,3,1,""],position:[12,2,1,""]},"s100py.s100.S100File":{PRODUCT_SPECIFICATION:[13,2,1,""],__init__:[13,3,1,""]},"s100py.s100.S100Root":{__init__:[14,3,1,""],epoch:[14,2,1,""],geographic_identifier:[14,2,1,""],horizontal_datum_reference:[14,2,1,""],horizontal_datum_value:[14,2,1,""],issue_date:[14,2,1,""],issue_time:[14,2,1,""],meta_features:[14,2,1,""],metadata:[14,2,1,""],product_specification:[14,2,1,""],vertical_datum:[14,2,1,""]},"s100py.s100.S100_FeatureContainer":{__init__:[15,3,1,""],axis_names:[15,2,1,""],common_point_rule:[15,2,1,""],data_coding_format:[15,2,1,""],dimension:[15,2,1,""],horizontal_position_uncertainty:[15,2,1,""],num_instances:[15,2,1,""],sequencing_rule_scan_direction:[15,2,1,""],time_uncertainty:[15,2,1,""],vertical_uncertainty:[15,2,1,""]},"s100py.s100.SEQUENCING_RULE_TYPE":{CantorDiagonal:[16,2,1,""],Hilbert:[16,2,1,""],Morton:[16,2,1,""],boustrophedonic:[16,2,1,""],linear:[16,2,1,""],spiral:[16,2,1,""]},"s100py.s100.SequenceRule":{__init__:[17,3,1,""],scan_direction:[17,2,1,""],type:[17,2,1,""]},"s100py.s100.VERTICAL_DATUM":{HAT:[18,2,1,""],LAT:[18,2,1,""],LW:[18,2,1,""],MHHW:[18,2,1,""],MHW:[18,2,1,""],MHWS:[18,2,1,""],MLLW:[18,2,1,""],MLW:[18,2,1,""],MLWS:[18,2,1,""],MSL:[18,2,1,""],approximateLowestAstronomicalTide:[18,2,1,""],approximateMeanLowWater:[18,2,1,""],approximateMeanLowWaterSprings:[18,2,1,""],approximateMeanLowerLowWater:[18,2,1,""],approximateMeanSeaLevel:[18,2,1,""],equinoctialSpringLowWater:[18,2,1,""],highWater:[18,2,1,""],highWaterSprings:[18,2,1,""],higherHighWaterLargeTide:[18,2,1,""],highestAstronomicalTide:[18,2,1,""],indianSpringLowWater:[18,2,1,""],internationalGreatLakesDatum1985:[18,2,1,""],localDatum:[18,2,1,""],lowWater:[18,2,1,""],lowWaterSprings:[18,2,1,""],lowerLowWaterLargeTide:[18,2,1,""],lowestAstronomicalTide:[18,2,1,""],lowestLowWater:[18,2,1,""],lowestLowWaterSprings:[18,2,1,""],meanHighWater:[18,2,1,""],meanHighWaterSprings:[18,2,1,""],meanHigherHighWater:[18,2,1,""],meanLowWater:[18,2,1,""],meanLowWaterSprings:[18,2,1,""],meanLowerLowWater:[18,2,1,""],meanLowerLowWaterSprings:[18,2,1,""],meanSeaLevel:[18,2,1,""],meanWaterLevel:[18,2,1,""],nearlyHighestHighWater:[18,2,1,""],nearlyLowestLowWater:[18,2,1,""]},"s100py.s100.VertexPoint":{__init__:[19,3,1,""],geometry:[19,2,1,""],value:[19,2,1,""]},"s100py.s102":{BathymetryContainer:[20,1,1,""],BathymetryCoverage:[21,1,1,""],BathymetryCoverage_dataset:[22,1,1,""],BathymetryCoverages_List:[23,1,1,""],BathymetryFeatureInstance:[24,1,1,""],BathymetryGroup_List:[25,1,1,""],BathymetryValues:[26,1,1,""],DiscoveryMetadata:[27,1,1,""],FeatureCodes:[28,1,1,""],FeatureInformation_dataset:[29,1,1,""],S102File:[30,1,1,""],S102Root:[31,1,1,""],S102_MetadataList_base:[32,1,1,""],SurfaceCorrectionValues:[33,1,1,""],TilingScheme:[34,1,1,""],TrackingListContainer:[35,1,1,""],TrackingListCoverage:[36,1,1,""],TrackingListCoverage_dataset:[37,1,1,""],TrackingListCoverages_List:[38,1,1,""],TrackingListGroup_List:[39,1,1,""],TrackingListSet_List:[40,1,1,""],TrackingListValues:[41,1,1,""],TrackingListValues_List:[42,1,1,""]},"s100py.s102.BathymetryContainer":{__init__:[20,3,1,""],bathymetry_coverage:[20,2,1,""]},"s100py.s102.BathymetryCoverage":{__init__:[21,3,1,""],axis_names:[21,2,1,""],dimension:[21,2,1,""],extent:[21,2,1,""],maximimum_display_scale:[21,2,1,""],maximum_depth:[21,2,1,""],maximum_uncertainty:[21,2,1,""],minimum_depth:[21,2,1,""],minimum_display_scale:[21,2,1,""],minimum_uncertainty:[21,2,1,""],offset_vectors:[21,2,1,""],origin:[21,2,1,""],sequencing_rule:[21,2,1,""],start_sequence:[21,2,1,""],values:[21,2,1,""],write_format_str:[21,2,1,""]},"s100py.s102.BathymetryCoverage_dataset":{__init__:[22,3,1,""],metadata_name:[22,2,1,""]},"s100py.s102.BathymetryCoverages_List":{__init__:[23,3,1,""],metadata_name:[23,2,1,""]},"s100py.s102.BathymetryFeatureInstance":{__init__:[24,3,1,""],bathymetry_group:[24,2,1,""]},"s100py.s102.BathymetryGroup_List":{__init__:[25,3,1,""],metadata_name:[25,2,1,""]},"s100py.s102.BathymetryValues":{__init__:[26,3,1,""],depth:[26,2,1,""],get_write_order:[26,3,1,""],metadata_name:[26,2,1,""],read:[26,3,1,""],uncertainty:[26,2,1,""],write:[26,3,1,""]},"s100py.s102.DiscoveryMetadata":{__init__:[27,3,1,""]},"s100py.s102.FeatureCodes":{__init__:[28,3,1,""],bathymetry_coverage_dataset:[28,2,1,""],feature_code:[28,2,1,""],feature_name:[28,2,1,""],tracking_list_coverage:[28,2,1,""]},"s100py.s102.FeatureInformation_dataset":{__init__:[29,3,1,""]},"s100py.s102.S102File":{PRODUCT_SPECIFICATION:[30,2,1,""],__init__:[30,3,1,""]},"s100py.s102.S102Root":{__init__:[31,3,1,""],bathymetry_coverage:[31,2,1,""],feature_information:[31,2,1,""],tracking_list_coverage:[31,2,1,""]},"s100py.s102.S102_MetadataList_base":{__init__:[32,3,1,""],write_format_str:[32,2,1,""]},"s100py.s102.SurfaceCorrectionValues":{__init__:[33,3,1,""]},"s100py.s102.TilingScheme":{__init__:[34,3,1,""]},"s100py.s102.TrackingListContainer":{__init__:[35,3,1,""],tracking_list_coverage:[35,2,1,""]},"s100py.s102.TrackingListCoverage":{__init__:[36,3,1,""],common_point_rule:[36,2,1,""],domain_extent:[36,2,1,""],set:[36,2,1,""],write_format_str:[36,2,1,""]},"s100py.s102.TrackingListCoverage_dataset":{__init__:[37,3,1,""],metadata_name:[37,2,1,""]},"s100py.s102.TrackingListCoverages_List":{__init__:[38,3,1,""],metadata_name:[38,2,1,""]},"s100py.s102.TrackingListGroup_List":{__init__:[39,3,1,""],metadata_name:[39,2,1,""]},"s100py.s102.TrackingListSet_List":{__init__:[40,3,1,""],metadata_name:[40,2,1,""]},"s100py.s102.TrackingListValues":{__init__:[41,3,1,""],list_series:[41,2,1,""],track_code:[41,2,1,""]},"s100py.s102.TrackingListValues_List":{__init__:[42,3,1,""],metadata_name:[42,2,1,""]},"s100py.s1xx":{S1XXFile:[43,1,1,""],S1XX_Attributes_base:[44,1,1,""],S1XX_Dataset_base:[45,1,1,""],S1XX_MetadataList_base:[46,1,1,""],S1XX_WritesOwnGroup_base:[47,1,1,""]},"s100py.s1xx.S1XXFile":{__init__:[43,3,1,""],create_empty_metadata:[43,3,1,""],depth_keys:[43,2,1,""],get_depth_dataset:[43,3,1,""],get_depths:[43,3,1,""],group_level_keys:[43,2,1,""],print_depth_attributes:[43,3,1,""],print_overview:[43,3,1,""],read:[43,3,1,""],second_level_keys:[43,2,1,""],show_keys:[43,3,1,""],top_level_keys:[43,2,1,""],tracking_list_group_level:[43,2,1,""],tracking_list_second_level:[43,2,1,""],tracking_list_top_level:[43,2,1,""],value_level_keys:[43,2,1,""],write:[43,3,1,""]},"s100py.s1xx.S1XX_Attributes_base":{S1XX_version:[44,2,1,""],__init__:[44,3,1,""],add_metadata:[44,3,1,""],get_all_keys:[44,3,1,""],get_metadata:[44,3,1,""],get_s1xx_attr:[44,3,1,""],get_standard_keys:[44,3,1,""],get_standard_list_properties:[44,3,1,""],get_standard_properties:[44,3,1,""],get_standard_properties_mapping:[44,3,1,""],get_write_order:[44,3,1,""],initialize_properties:[44,3,1,""],read:[44,3,1,""],read_hdf5_attributes:[44,3,1,""],set_enum_attribute:[44,3,1,""],set_s1xx_attr:[44,3,1,""],write:[44,3,1,""],write_as_xml:[44,3,1,""]},"s100py.s1xx.S1XX_Dataset_base":{__init__:[45,3,1,""],append_new_item:[45,3,1,""],metadata_name:[45,2,1,""],read:[45,3,1,""],write:[45,3,1,""],write_as_xml:[45,3,1,""]},"s100py.s1xx.S1XX_MetadataList_base":{__init__:[46,3,1,""],append_new_item:[46,3,1,""],metadata_name:[46,2,1,""],read:[46,3,1,""],read_re_pattern:[46,2,1,""],write:[46,3,1,""],write_as_xml:[46,3,1,""],write_format_str:[46,2,1,""]},"s100py.s1xx.S1XX_WritesOwnGroup_base":{__init__:[47,3,1,""]},s100py:{s100:[49,0,0,"-"],s102:[51,0,0,"-"],s1xx:[52,0,0,"-"]}},objnames:{"0":["py","module","Python module"],"1":["py","class","Python class"],"2":["py","attribute","Python attribute"],"3":["py","method","Python method"]},objtypes:{"0":"py:module","1":"py:class","2":"py:attribute","3":"py:method"},terms:{"":[8,13,17,30,50,53],"02d":36,"03d":[21,32,46],"10c":[3,5,7,14,20,53],"abstract":7,"boolean":8,"byte":[13,30,43],"case":[3,48,53],"class":[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,50,53],"default":[13,17,21,30,36,43,48,53],"enum":[0,1,11,14,16,18,49,53],"final":[50,53],"float":[3,4,6,7,15,21,53],"function":[11,44,48,50,53],"import":[26,44,50,53],"int":[2,4,5,6,13,14,15,21,30,41,48,53],"long":18,"new":[13,30,43,48,53],"return":[2,3,4,5,6,7,8,9,10,12,14,15,17,18,19,20,21,22,23,24,25,26,28,31,35,36,37,38,39,40,41,42,44,45,46,48,50,53],"super":44,"true":[13,30,43,50],"try":53,"while":[3,10,13,17,18,29,30,43],But:53,For:[13,17,30,43,44,48,50,53],IDE:53,That:[8,53],The:[3,5,7,8,9,10,12,13,17,19,20,21,24,26,29,30,31,35,36,41,43,44,45,46,47,50,53],There:[8,19,21,50,53],These:[19,36,50,53],Use:53,Using:50,__init__:[2,3,4,5,6,7,8,9,10,12,13,14,15,17,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,53],__version__:[44,53],_attribut:[48,50,53],_attribute_nam:[44,48],_creat:48,_hl:[43,49,51,52],_nnn:50,_type:48,abc:[44,49,51,52],about:[43,52],abov:[48,50,53],accept:53,access:[45,50,53],accord:[13,30,43],accur:[2,3,4,5,6,7,8,9,10,12,14,15,17,19,20,21,22,23,24,25,26,27,28,29,31,32,33,34,35,36,37,38,39,40,41,42,44,45,46,47],act:[23,46,53],actual:[3,29,53],add:[43,50,53],add_metadata:44,addit:[13,30,43,53],aforement:50,alias:18,all:[0,13,17,30,43,44,50,53],allow:[13,17,21,30,36,43,50,53],along:11,alreadi:53,also:[7,21,36,50,53],alt:48,alwai:[13,30,43],amd:50,ani:[11,26,45,50,53],anoth:53,aovvaw07qesny5urachwio1e4alu:10,api:[50,52],appear:50,append_new_item:[45,46,50],appli:[11,47],applic:[13,30,43,48],appropri:53,approxim:[13,30,43],approximatelowestastronomicaltid:18,approximatemeanlowerlowwat:18,approximatemeanlowwat:18,approximatemeanlowwaterspr:18,approximatemeansealevel:18,arbitrari:[36,50,53],arc:11,area:[8,11],arg:[13,22,23,25,27,29,30,32,37,38,39,40,42,43,45,46],arrai:[43,44,46,48,53],assign:[11,19,36],associ:[11,21,44,53],at_download:10,attr:[43,48],attr_float:53,attr_float_attribute_nam:53,attr_float_cr:53,attr_float_typ:53,attr_int:53,attr_int_attribute_nam:53,attr_int_cr:53,attr_int_typ:53,attr_str:53,attr_str_attribute_nam:53,attr_str_creat:53,attr_str_typ:53,attrfloat:53,attribut:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,28,30,31,32,35,36,37,38,39,40,41,42,43,44,45,46,48,50,53],attribute_nam:44,attrint:53,attrstr:53,atttribut:46,auto:53,autodetermin:44,automat:53,averag:0,axi:[9,17],axis_nam:[15,21],axisnam:9,barycentr:11,base:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,49,51,52,53],basic:[3,12,29,50,53],bathi:50,bathmetrycoverag:50,bathy_01:50,bathy_list:50,bathymetr:36,bathymetri:[20,24,31,41,50],bathymetry_coverag:[20,31,44,50],bathymetry_coverage_attribute_nam:50,bathymetry_coverage_cr:50,bathymetry_coverage_dataset:28,bathymetry_coverage_typ:50,bathymetry_group:24,bathymetrycontain:[50,51],bathymetrycoverag:[3,7,8,20,23,24,29,43,44,50,51],bathymetrycoverage_01:43,bathymetrycoverage_dataset:[28,50,51],bathymetrycoverages_list:[20,24,50,51],bathymetryfeatureinst:[50,51],bathymetrygroup:24,bathymetrygroup_list:[50,51],bathymetryvalu:[21,50,51],bathymetryvaluerecord:[],bathymetryvalueslist:[],becaus:[50,53],been:[13,30,43,50],befor:[13,30,43],begin:50,behav:53,being:[50,53],believ:28,belong:53,below:48,better:53,between:[11,13,30,43,53],bicub:11,bilinear:11,biquadrat:11,blank:[48,50,53],block:[13,30,43],bool:8,bottom:[48,53],bound:[7,8,13,30,43],boundari:[7,8],boustrophedon:16,box:[7,8],built:44,button:48,cach:[13,30,43],call:[44,47,50],camel:[3,48,53],camelcas:53,can:[10,13,18,30,43,47,50,53],cantordiagon:16,capit:17,care:53,carri:8,cell:[11,17,21],chanc:53,chang:48,charact:[15,21],characterstr:[17,41],chees:53,child:[20,50,53],children:50,chunk:[13,30,43],click:48,closedinterv:3,closur:3,code:[3,11,28,48,50,52,53],codelist:17,collect:[49,51,52],collis:[13,30,43],column:10,combin:47,commetn:[],common_point_rul:[15,36,49,50],commonpointrul:36,commonstateobject:[49,51,52],complex:50,compon:8,composit:[],compound:53,compund:53,comput:11,conform:[17,21],connect:11,consist:12,contain:[20,23,24,31,35,41,43,48,49,50,51,52,53],coord_valu:9,coordin:[2,9,10,36],coordvalu:9,core:[13,30,43],corner:21,correct:28,correspond:[8,19,41],could:[45,53],cover:8,coverag:[11,19,31,36,41,50],coveragesattribut:31,creat:[13,18,26,30,43,44,45,47,48,50,52,53],create_dataset:43,create_empty_metadata:43,create_group:43,creation:[13,30,43,48],ctrl:48,cubic:11,current:[13,30,31,43],cursor:48,curv:11,cv_discretepointcoverag:36,cv_geometryvaluepair:36,cv_gridcoordin:21,cv_gridenvelop:[7,8],cv_interpolationmethod:11,cv_sequencerul:[17,21],data:[3,8,10,12,13,17,19,20,21,24,29,30,35,43,45,46,47,48,50,53],data_coding_format:[15,49,50],data_grid:53,data_grid_attribute_nam:53,data_grid_cr:53,data_grid_typ:53,data_valu:53,data_value_attribute_nam:53,data_value_cr:53,data_value_typ:53,datagrid:53,datagroup:53,datagroup_01:53,datagroup_02:53,datagroup_nnn:53,datagroupobject:53,datanam:50,dataname_attribute_nam:50,dataname_cr:50,dataname_typ:50,dataset:[2,3,9,13,15,17,19,21,25,26,28,30,43,44,45,47,48,50,53],datasetwithnam:53,datasetwithnames_list:53,datatyp:[3,53],datavalu:53,date_time_of_first_record:5,date_time_of_last_record:5,datum:18,def:[48,53],defin:[7,13,30,36,41,43,48,53],definit:[],denot:3,depend:[13,30,43],depth:[3,19,25,26,29,36,41,43,45],depth_kei:43,depthcorrectiontyp:[],deriv:[19,47,50,53],derivedgroupfroms100:[],describ:[3,8,10,11,21,29,41,53],descript:48,desir:[13,30,43],detail:[13,30,43,52],determin:[13,21,30,31,43,50],diagram:50,dict:44,dictionari:[43,44,53],differ:[13,30,43,50],dimens:[2,9,15,21],direct:[9,17,21],directli:52,directposit:[12,21,49,50],discov:53,discoverymetadata:[50,51],discret:[11,19],disk:[13,30,43,53],display_nod:43,distanc:11,doc:[10,48,53],document:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,28,30,31,32,35,36,37,38,39,40,41,42,43,44,45,46,53],doe:48,doesn:53,doing:53,domain:11,domain_ext:36,done:53,dot:[50,53],driver:[13,30,43],due:[13,30,43],duplic:50,dure:[26,45],each:[9,19,21,25,36,50,53],earliest:[13,30,43],east:17,east_bound_longitud:[7,50],eastboundlongitud:[50,53],edit:[36,48],editor:48,either:47,element:[3,7,8,9,21,29,41,44],elev:43,empti:[13,30,43,48,50,53],encapsul:[50,53],encod:[50,53],encompass:8,end:[26,45,53],entri:[3,29],enum_typ:44,enumer:[0,1,11,16,17,48,53],envelop:10,epoch:[14,50],equat:53,equinoctialspringlowwat:18,error:28,establish:[7,8],etc:[3,18,23,29,44,46],etree_object:[44,45,46],even:53,eventu:[44,53],everyth:50,ex_ext:8,ex_geographicboundingbox:7,ex_geographicext:[7,8],exampl:[6,50],except:53,exist:[13,30,43,48],expect:50,explan:[13,30,43],express:48,extend:[50,52,53],extending_the_api:53,extens:[11,50],extent:[8,10,21],extent_type_cod:[8,50],extenttypecod:[8,50],extra:[26,45],fail:[13,30,43],fals:[2,3,4,5,6,7,8,9,10,12,14,15,17,19,20,21,24,26,28,31,33,34,35,36,41,44,47,50],famili:[50,52],faster:53,feat:45,featur:[3,5,11,17,29,31],feature_cod:28,feature_inform:[31,44],feature_information_cr:50,feature_nam:28,featurecod:[31,50,51],featureinform:[45,49,50],featureinformation_dataset:[22,37,50,51],featureinst:[5,6,49,50,51],featureinstance_format_2:[24,49,50,51],featureinstancetimeseri:[49,50],featurenam:28,fgdc:10,field:41,figur:[12,19,21,23,31,36,38],file:[10,13,18,26,30,43,44,45,46,48,49,50,51,52,53],fill:[48,53],fill_empti:[2,3,4,5,6,7,8,9,10,12,14,15,17,19,20,21,24,26,28,31,33,34,35,36,41,44,47],fill_valu:3,first:[21,48,50,53],fit:[13,30,43],flush:[13,30,43],follow:53,format:[6,50,53],four:53,frameword:10,framework:[48,53],friendli:53,from:[5,7,8,9,12,13,14,17,18,19,20,26,30,31,36,43,44,45,46,48,50,53],fulli:[13,30,43],further:[17,52],gener:[44,50],geographic_identifi:14,geographicboundingbox:[4,14,49,50,51,53],geographicext:[7,49,50,51],geometr:[11,12],geometri:[19,36],get:[44,50],get_all_kei:44,get_config:[13,30,43],get_depth:43,get_depth_dataset:43,get_metadata:44,get_s1xx_attr:44,get_standard_kei:44,get_standard_list_properti:44,get_standard_properti:[44,50],get_standard_properties_map:[44,50],get_write_ord:[26,44,53],gi_frameworkdatastandard_part3_elev:10,give:48,global:[13,30,43],gm_point:12,going:50,gotten:10,gov:10,grab:50,grid:[6,9,10,11,17,19,21,53],grid_origin_latitud:6,grid_origin_longitud:6,grid_origin_vert:6,grid_spacing_latitudin:6,grid_spacing_longitudin:6,grid_spacing_vert:6,gridcoordin:[21,49,50],gridenvelop:[21,49,50],gridoriginlatitud:50,gridoriginlongitud:50,group:[5,13,20,23,25,26,30,31,43,44,45,46,47,48,49,50,51,52,53],group_001:[43,53],group_01:53,group_02:53,group_f:44,group_level_kei:43,group_nnn:48,group_object:[26,44,45,46],groupf:[3,29],guid:[13,30,43],h5py:[13,26,30,43,44,45,46,49,50,51,52,53],h5t_native_float:3,handl:50,harder:53,has:[9,17,19,20,21,25,41,44,48,50,53],hash:[13,30,43],hat:18,have:[3,13,29,30,43,44,47,48,50,53],hdf5:[13,19,20,26,30,43,44,45,46,48,50,53],hdf:53,held:[10,25,50,53],help:[2,3,4,5,6,7,8,9,10,12,14,15,17,19,20,21,22,23,24,25,26,27,28,29,31,32,33,34,35,36,37,38,39,40,41,42,44,45,46,47,53],helper:53,here:[23,48,53],high:[0,10],higherhighwaterlargetid:18,highestastronomicaltid:18,highlight:48,highwat:18,highwaterspr:18,hilbert:16,hint:53,hit:[48,53],hlobject:[49,51,52],hold:[9,19,20,53],holder:50,horizont:36,horizontal_datum_refer:14,horizontal_datum_valu:14,horizontal_position_uncertainti:15,housekeep:53,how:[13,21,30,43,48,53],html:[],http:10,human:53,hyperlink:48,ideal:[13,30,43],identifi:[21,36],iho:[13,30],illeg:53,implement:[44,52,53],inclus:[13,30,43],increas:[13,30,43],indent:[26,43,44,45,46],index:[41,45],indianspringlowwat:18,indic:[7,8,10,13,17,19,30,43],inform:[8,31],inherit:50,initi:[2,3,4,5,6,7,8,9,10,12,14,15,17,19,20,21,22,23,24,25,26,27,28,29,31,32,33,34,35,36,37,38,39,40,41,42,44,45,46,47,50],initialize_properti:[44,50],input:53,insead:[],insert:43,insid:[50,53],instanc:[5,45,53],instance_chunk:4,instant:53,instead:[10,28,53],integ:[9,10,41,53],internationalgreatlakesdatum1985:18,interpol:11,interpolation_typ:[49,50],isinst:53,iso:[7,8,10,11,12,17,19,21,36],issue_d:14,issue_tim:14,item:[19,26,44,48,50],iter:[49,51,52],its:[47,50],itself:[50,53],jsut:53,just:53,keep:53,kei:[26,44],keyerror:50,keyword:[13,30,43],know:53,known:[11,17],kwarg:27,kywrd:[13,30,43],lai:53,last:[50,53],lat:18,latest:[13,30,43],latitid:50,latitud:[17,50],layer:17,least:[13,30,43],left:[10,21],legal:[13,30,43],length:11,let:[50,53],level:[3,29,50],librari:[13,30,43],libver:[13,30,43],like:[13,23,28,30,43,44,46,53],limit:53,line:[11,48,50],lineag:41,linear:[11,16,17,21],list:[3,7,8,17,19,20,23,24,25,26,29,31,35,36,41,44,45,46,48,50,53],list_seri:41,listseri:[19,41],liter:11,littl:50,live:[48,53],livetempl:48,localdatum:18,locat:[19,53],logic:47,longitud:17,look:53,lost:[11,53],lostarea:11,lot:53,low:[0,10],lower:[3,10,13,21,30,43],lowercas:17,lowerlowwaterlargetid:18,lowestastronomicaltid:18,lowestlowwat:18,lowestlowwaterspr:18,lowwat:18,lowwaterspr:18,lru:[13,30,43],made:53,mai:[8,41,53],make:[43,46,47,48,53],manag:50,map:[17,49,51,52],mappinghdf5:[49,51,52],match:[3,19,29,31,53],matrix:10,maximimum_display_scal:21,maximum:[13,30,43],maximum_depth:21,maximum_uncertainti:21,mean:[13,30,43],meanhigherhighwat:18,meanhighwat:18,meanhighwaterspr:18,meanlowerlowwat:18,meanlowerlowwaterspr:18,meanlowwat:18,meanlowwaterspr:18,meansealevel:18,meanwaterlevel:18,measur:3,member:[48,50,53],memori:[13,30,43],meta_featur:14,metaclass:8,metadata:[3,7,8,14,29,36,41,43,50],metadata_nam:[22,23,25,26,37,38,39,40,42,45,46,53],metadata_typ:53,method:[11,26,43,44,45,46,50],metr:3,mhhw:18,mhw:18,miniconda36:50,minimum_depth:21,minimum_display_scal:21,minimum_uncertainti:21,miss:3,mllw:18,mlw:18,mode:[13,30,43],model:36,modifi:[48,52],modul:50,monti:[48,53],more:[36,50,53],morton:[16,17,21],most:[28,50],mpio:[13,30,43],msl:18,much:53,multi:53,multipl:[44,47,48,50,53],must:[13,30,43,44,53],mutablemap:[49,51,52],mutablemappinghdf5:[49,51,52],myfirstobject:53,myloc:53,mylocationgroup:53,myobject:53,name:[3,13,17,18,20,26,28,29,30,31,43,44,45,46,47,48,50,53],name_nnn:50,name_of_data:[48,53],name_of_data_attribute_nam:[48,53],name_of_data_cr:53,name_of_data_typ:[48,53],nameofdata:53,natur:53,navo:28,ndarrai:[2,9,15,17,19,21,26,28],nearest:11,nearestneighbor:11,nearlyhighesthighwat:18,nearlylowestlowwat:18,necessari:36,need:[47,53],newer:53,next:[50,53],nnn:[23,25,46],noaa:50,nomenclatur:50,non:[13,30,43,53],none:[13,26,30,43,44],north:[17,50],north_bound_latitud:[7,50],northboundlatitud:50,northboundlongitud:53,notat:50,note:[13,18,21,30,43,46,50],notic:53,nov:44,now:[50,53],num_grp:4,num_inst:15,num_points_latitudin:6,num_points_longitudin:6,num_points_vert:6,number:[9,13,30,41,43,50,53],number_of_tim:5,numpi:[44,48,53],obj:43,object:[11,12,13,18,20,24,26,30,35,43,44,45,50,53],occur:44,occurr:[48,50,53],off:20,offici:18,offset:[9,10,21],offset_vector:21,offsetvector:[10,21],often:[13,30,43],older:53,omit:50,onc:[13,30,43,53],one:[8,9,12,36,48,50,52,53],onli:[8,11,12,13,17,21,30,43,48,52,53],open:[13,30,43],opt:[22,23,25,29,32,37,38,39,40,42,45,46],option:[8,13,17,21,30,36,41,43,50],order:[9,13,17,21,26,30,43,44,45],ordereddict:50,org:[],organ:36,origin:[9,10,21],other:[13,17,21,23,30,43,53],otherwis:[13,26,28,30,43,44],our:[50,53],out:[26,45,53],output:53,overrid:[26,41,44],overridden:[19,44],overwrit:50,own:[47,50],packag:50,pad:50,pair:[11,36],parallel:[3,29],paramet:[11,26,44,45,46],parent:47,part:[46,53],pass:[13,30,43],past:48,pattern:53,penal:[13,30,43],pep8:[50,53],per:[13,30,43],perform:[13,30,43],pertain:50,piec:53,place:[48,50],plain:[3,53],plu:48,point:[11,17,19,21,36,49,50],polici:[13,30,43],polygon:[8,11],popul:50,portion:8,posit:[11,12,17],possibl:20,pre:47,predict:47,preempt:[13,30,43],preemption:[13,30,43],prescrib:50,present:8,press:48,pretend:53,primari:[43,50],prime:[13,30,43],print:50,print_depth_attribut:43,print_overview:43,privat:53,product:50,product_specif:[13,14,30],project:10,properti:[44,48,50,53],provid:53,publish:44,put:[48,50],pycharm:[48,53],pydrotrunk:50,python3:50,python:[46,48,50,53],quad:[17,21],quadrat:11,quadrilater:11,quit:53,rais:53,raw:[13,30,43],rdcc_nbyte:[13,30,43],rdcc_nslot:[13,30,43],rdcc_w0:[13,30,43],read:[10,13,26,30,43,44,45,46,48,50,53],read_hdf5_attribut:44,read_re_pattern:46,readonli:[13,30,43],real:10,realli:[46,50],reason:41,recal:53,recent:50,recommend:[13,30,43],record:[17,19,21,36,46],rectangular:53,recurs:50,reduc:[13,30,43],refactor:50,refer:50,referenc:[19,50],regular:[17,21],relat:50,relationship:[],remov:28,repreent:46,repres:[36,46],requir:[13,21,30,43,53],right:[10,20,48],root:[13,20,30,31,43,44,50,53],row:10,rtype:[3,20,24,26,35,44],rule:[13,30,43],run:53,s100:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,24,31,33,35,36,46,48,50,51,52,53],s100_cv_interpolationmethod:11,s100_featurecontain:[20,35,49,50,51],s100_grid:[],s100_point:36,s100_pointcoverag:36,s100file:[49,50],s100py:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,53],s100root:[31,49,50,51],s102:[3,6,17,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,44,45,46,48,50],s102_bathymetrycoverag:43,s102_bathymetrycoverage_01:43,s102_bathymetryvalu:[],s102_bathymetryvaluerecord:[],s102_dataidentif:[],s102_elev:43,s102_grid:43,s102_grid_01:43,s102_metadatalist_bas:[20,23,24,25,35,36,38,39,40,42,50,51],s102_trackinglistvalu:19,s102file:[50,51],s102root:[50,51],s1xx:[2,3,8,9,10,12,13,15,17,19,21,26,27,28,29,30,32,34,36,43,44,45,46,47,49,50,51,53],s1xx_attributes_bas:[2,3,8,9,10,12,15,17,19,21,27,28,31,34,36,47,48,49,50,51,52,53],s1xx_dataset_bas:[29,44,48,50,51,52,53],s1xx_metadatalist_bas:[32,44,48,50,51,52,53],s1xx_name:44,s1xx_sequenc:53,s1xx_version:44,s1xx_writesowngroup_bas:[26,45,46,50,51,52],s1xxattribute_bas:53,s1xxfile:[13,30,49,50,51,52],s999:53,safe:[13,30,43],sai:[48,50,53],same:[9,10,13,30,43,53],sampl:[50,53],scan:17,scan_direct:17,scandirect:17,scheme:48,sec2:[13,30,43],sec:[10,36],second:53,second_level_kei:43,sect:28,section:21,see:[2,3,4,5,6,7,8,9,10,12,13,14,15,17,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,50,52],seem:10,segment:11,select:[13,30,43,48],self:[2,3,4,5,6,7,8,9,10,12,14,15,17,19,20,21,22,23,24,25,26,27,28,29,31,32,33,34,35,36,37,38,39,40,41,42,44,45,46,47,48,50,53],sequenc:[9,15,17,19,21],sequencerul:[21,49,50],sequencetyp:17,sequencing_rul:21,sequencing_rule_scan_direct:15,sequencing_rule_typ:[49,50],sequencingrul:21,set:[8,13,17,23,30,36,43,44,48,50],set_enum_attribut:[44,48,53],set_s1xx_attr:44,setter:[48,53],shall:[9,17,19,21],shell:53,shorthand:18,should:[13,19,28,30,43,45,50,52,53],show:[50,53],show_kei:43,side:48,signatur:[2,3,4,5,6,7,8,9,10,12,14,15,17,19,20,21,22,23,24,25,26,27,28,29,31,32,33,34,35,36,37,38,39,40,41,42,44,45,46,47],similar:[48,53],simpl:[17,21,53],sinc:10,singl:9,site:50,size:[10,13,17,21,30,43,49,51,52],skeleton:50,slightli:[13,30,43],slot:[13,30,43],snake:53,snakecas:48,softwar:28,some:[45,50,53],someotherattribut:[],sound:36,south:17,south_bound_latitud:7,southboundlatitud:53,space:21,spam:53,spatial:[7,8,36],spec:[5,14,20,45,46,50,53],special:47,specif:[9,50,52],specifi:[10,17,21,26,45,47,48,53],speed:48,spiral:16,standard:[10,17,21,53],start:53,start_sequ:[6,21],startsequ:21,stdio:[13,30,43],still:[13,30,43,53],store:[3,29,44,45,53],str:[3,4,5,6,14,15,17,22,23,25,26,36,37,38,39,40,41,42,45,46,48,53],straight:53,strategi:[13,30,43],strictli:[13,30,43],string:[3,13,15,17,18,21,30,41,43,53],structur:50,style:53,sub:[44,47,48,53],subgroup:[44,53],subtyp:[7,8],summari:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,28,30,31,32,35,36,37,38,39,40,41,42,43,44,45,46],suppli:[46,53],support:53,surfacecorrectionvalu:[41,50,51],swmr:[13,30,43],tab:53,tabl:[3,5,7,14,20,21,23,28,29,34,35,38,53],take:[46,48,53],taken:12,tell:48,temp:50,templat:[48,53],test:50,text:[3,41,48],than:[13,30,43,50],thei:[23,50,52,53],them:[45,50,53],themselv:53,therefor:36,therein:50,thi:[3,10,13,17,20,21,23,25,26,28,29,30,41,43,44,45,46,48,50,52,53],thiessen:11,thing:[44,45,47,53],think:10,three:[50,53],thumb:[13,30,43],tilingschem:[50,51],time:[13,30,43],time_record_interv:5,time_uncertainti:15,tin:11,too:[26,44,45,46],top:[3,29,50],top_level_kei:43,toplevelattribut:[],toplevelgroup:[],total:[13,30,43],traceback:50,track:[7,8,13,19,30,31,35,43],track_cod:41,track_ord:[13,30,43],trackcod:41,tracking_list_coverag:[28,31,35,44],tracking_list_group_level:43,tracking_list_second_level:43,tracking_list_top_level:43,trackinglistcontain:[50,51],trackinglistcoverag:[3,29,35,43,44,50,51],trackinglistcoverage_dataset:[28,50,51],trackinglistcoverages_list:[35,50,51],trackinglistgroup_list:[50,51],trackinglistset_list:[50,51],trackinglistvalu:[36,50,51],trackinglistvalues_list:[36,50,51],trail:50,treat:[13,30,43],tree:[17,21,53],truncat:[13,30,43],tupl:44,two:[11,19,31,50,53],type:[2,3,4,5,6,7,8,9,10,11,12,14,15,17,19,20,21,22,23,24,25,26,27,28,29,31,32,33,34,35,36,37,38,39,40,41,42,44,45,46,47,48,50,53],typehint:53,uml:46,uncertainti:[3,19,25,26,41,45],under:[3,13,29,30,43,48],underscor:[50,53],union:[2,9,15,17,19,21,26,28,48,53],unit:3,unit_of_measur:3,unnecessari:10,unord:[26,45],upper:[3,10],use:[8,13,30,43,44,48],used:[3,6,7,8,13,17,21,29,30,43,45,50,52,53],user:[13,30,41,43,53],userblock:[13,30,43],uses:[50,53],usg:10,using:[10,11,26,45,46,48,50,52,53],utm_zon:53,utm_zone_attribute_nam:53,utm_zone_cr:53,utm_zone_typ:53,utmzon:53,val:[44,48,53],valid:53,valu:[3,8,9,10,11,13,17,19,21,26,30,36,41,43,44,45,48,50,53],value_level_kei:43,variabl:[48,53],vector:21,veri:48,version:[13,30,43,44],vertexpoint:[33,49,50,51],vertic:18,vertical_datum:[14,49,50],vertical_extent_maximum_z:4,vertical_extent_minimum_z:4,vertical_uncertainti:15,verticaluncertaintytyp:[],via:[20,24,35],wai:50,want:46,weight:[13,30,43],well:[23,36],were:19,west:17,west_bound_longitud:[7,50],westboundlongitud:[44,53],what:[48,50,53],when:[8,13,30,43,50,53],where:[8,50],whether:8,which:[10,13,17,19,20,30,31,41,43,44,46,48,50,53],within:[3,7,8,11,17,29],without:47,word:48,work:[44,53],world:10,would:[10,45,50,53],wrap:53,write:[13,26,30,43,44,45,46,48,53],write_as_xml:[44,45,46],write_format_str:[21,32,36,46],writer:47,written:[13,18,30,43,44,50],www:10,xml:44,yes:50,you:[13,30,43,50,53],your:[13,30,43,48],zero:[48,50,53],zone:53},titles:["COMMON_POINT_RULE","DATA_CODING_FORMAT","DirectPosition","FeatureInformation","FeatureInstance","FeatureInstanceTimeSeries","FeatureInstance_Format_2","GeographicBoundingBox","GeographicExtent","GridCoordinate","GridEnvelope","INTERPOLATION_TYPE","Point","S100File","S100Root","S100_FeatureContainer","SEQUENCING_RULE_TYPE","SequenceRule","VERTICAL_DATUM","VertexPoint","BathymetryContainer","BathymetryCoverage","BathymetryCoverage_dataset","BathymetryCoverages_List","BathymetryFeatureInstance","BathymetryGroup_List","BathymetryValues","DiscoveryMetadata","FeatureCodes","FeatureInformation_dataset","S102File","S102Root","S102_MetadataList_base","SurfaceCorrectionValues","TilingScheme","TrackingListContainer","TrackingListCoverage","TrackingListCoverage_dataset","TrackingListCoverages_List","TrackingListGroup_List","TrackingListSet_List","TrackingListValues","TrackingListValues_List","S1XXFile","S1XX_Attributes_base","S1XX_Dataset_base","S1XX_MetadataList_base","S1XX_WritesOwnGroup_base","Extending the API","s100py.s100 Module","s100py","s100py.s102 Module","s100py.s1xx Module","An Example API"],titleterms:{"class":[49,51,52],"new":[],api:[48,53],bathymetrycontain:20,bathymetrycoverag:21,bathymetrycoverage_dataset:22,bathymetrycoverages_list:23,bathymetryfeatureinst:24,bathymetrygroup_list:25,bathymetryvalu:26,bathymetryvaluerecord:[],bathymetryvalueslist:[],common_point_rul:0,data_coding_format:1,diagram:[49,51,52],directposit:2,discoverymetadata:27,exampl:53,extend:48,featurecod:28,featureinform:3,featureinformation_dataset:29,featureinst:4,featureinstance_format_2:6,featureinstancetimeseri:5,geographicboundingbox:7,geographicext:8,gridcoordin:9,gridenvelop:10,inherit:[49,51,52],interpolation_typ:11,modul:[49,51,52],point:12,s100:49,s100_featurecontain:15,s100file:13,s100py:[49,50,51,52],s100root:14,s102:51,s102_metadatalist_bas:32,s102file:30,s102root:31,s1xx:52,s1xx_attributes_bas:44,s1xx_dataset_bas:45,s1xx_metadatalist_bas:46,s1xx_writesowngroup_bas:47,s1xxfile:43,sequencerul:17,sequencing_rule_typ:16,surfacecorrectionvalu:33,tilingschem:34,trackinglistcontain:35,trackinglistcoverag:36,trackinglistcoverage_dataset:37,trackinglistcoverages_list:38,trackinglistgroup_list:39,trackinglistset_list:40,trackinglistvalu:41,trackinglistvalues_list:42,vertexpoint:19,vertical_datum:18}})
